InplaceArrays.jl

InplaceArrays.jl

Documentation for InplaceArrays.jl

The Functor interface

Often, it is needed to implement functions that need some scratch data (e.g., pre-allocating the output). The question is, where to store this data? There are three main answers to this question: 1) store the data in the function object as part of its state, 2) allocate the scratch data each time the operation is performed, and 3) the user allocates and passes the scratch data when needed. Clearly, 1) it is not save if several calls to the operation are using the same scratch data (e.g., multi-threading). 2) is save, but it can be inefficient if the operation is performed at low granularity. 3) is both save and efficient, but requires some extra work by the user.

In Gridap, we adopt the 3rd option. In order to unify the interfaces of functions using this approach, we introduce the Functor interface. Any type is referred to as a Functor if it implements the following interface. We rely in duck typing here. There is not an abstract type representing a functor.

InplaceArrays.Functors.evaluate_functor! — Function.

y = evaluate_functor!(cache,f,x...)

Evaluates the functor f at the arguments x... using the scratch data provided in the given cache object. The cache object is built with the functor_cache function using arguments of the same type as in x... In general, the returned value y can share some part of its state with the cache object. If the result of two or more invocations of this function need to be accessed simultaneously (e.g., in multi-threading), create and use various cache objects (e.g., one cache per thread).

source

InplaceArrays.Functors.functor_cache — Function.

cache = functor_cache(f,x...)

Returns the cache needed to evaluate functor f with arguments of the same type as the objects in x....

source

Default implementations

We provide some default (obvious) implementations of this interface so that Function, Number, and AbstractArray objects behave like functors.

Examples

Calling the + function via the functor interface.

julia> cache = functor_cache(+,0,0)

julia> evaluate_functor!(cache,+,1,2)
3

julia> evaluate_functor!(cache,+,-1,10)
9

Number and AbstractArray objects behave like "constant" functors.

julia> a = 2.0
2.0

julia> cache = functor_cache(a,0)

julia> evaluate_functor!(cache,a,1)
2.0

julia> evaluate_functor!(cache,a,2)
2.0

julia> evaluate_functor!(cache,a,3)
2.0

Evaluating a functor without cache

InplaceArrays.Functors.evaluate_functor — Function.

evaluate_functor(f,x...)

Evaluate the fuctor f at the arguments x... by creating a temporary cache internally. This functions is equivalent to

cache = functor_cache(f,x...)
evaluate_functor!(cache,f,x...)

source

Working with several functors at once

InplaceArrays.Functors.evaluate_functors! — Function.

evaluate_functors!(caches::Tuple,fs::Tuple,x...) -> Tuple

Evaluates the functors in the tuple fs at the arguments x... by using the corresponding cache objects in the tuple caches. The result is also a tuple containing the result for each functor in fs.

source

InplaceArrays.Functors.functor_caches — Function.

functor_caches(fs::Tuple,x...) -> Tuple

Returns a tuple with the cache corresponding to each functor in fs for the arguments x....

source

Broadcasting

InplaceArrays.Functors.bcast — Function.

bcast(f::Function)

Returns a functor object that represents the "boradcasted" version of the given function f.

Examples

julia> op = bcast(*)
InplaceArrays.Functors.BCasted{typeof(*)}(*)

julia> x = ones(2,3)
2×3 Array{Float64,2}:
 1.0  1.0  1.0
 1.0  1.0  1.0

julia> y = 2
2

julia> evaluate_functor(op,x,y)
2×3 CachedArray{Float64,2,Array{Float64,2}}:
 2.0  2.0  2.0
 2.0  2.0  2.0

source

Composition

InplaceArrays.Functors.compose_functors — Function.

c = compose_functors(g,fs...)

Returns an object c representing the "composition" of functor g with several functors fs. The resulting object c is such that

evaluate_functor(c,x...)

is equivalent to

fxs = evaluate_functors(fs,x...)
evaluate_functor(g,fxs...)

source

Extended AbstractArray interface

When implementing new array types, we encounter a similar problem than when implementing some functions : It can be needed some scratch data (e.g., allocating the output), when recovering an item from an array (typically if the array elements are mutable or non-isbits objects, e.g., for "lazy" array of arrays). Here, we adopt the same solution as for functors: the user provides the scratch data. However, the Julia array interface does not support this approach. When calling a[i], in order to get the element with index i in array a, there is no extra argument for the scratch data. In order to circumvent this problem, we add new methods to the AbstractArray interface of Julia. We provide default implementations to the new methods, so that any AbstractArray can be used with the extended interface. The most important among the new methods is getindex!, which allows to recover an item in the array by passing some scratch data. The new mehtods are listed below.

InplaceArrays.Arrays.getindex! — Function.

getindex!(cache,a::AbstractArray,i...)

Returns the item of the array a associated with index i by (possibly) using the scratch data passed in the cache object.

It defaults to

getindex!(cache,a::AbstractArray,i...) = a[i...]

The cache object is constructed with the array_cache function.

Examples

julia> a = collect(1:4)
4-element Array{Int64,1}:
 1
 2
 3
 4

julia> cache = array_cache(a)

julia> getindex!(cache,a,2)
2

julia> getindex!(cache,a,4)
4

In this example, using the extended interface provides little benefit, but for new array types that need scratch data, efficient implementations of getindex! can make a performance difference by avoiding low granularity allocations.

source

InplaceArrays.Arrays.array_cache — Function.

array_cache(a::AbstractArray)

Returns a cache object to be used in the getindex! function. It defaults to

array_cache(a::T) where T = nothing

for types T such that uses_hash(T) == Val(false), and

function array_cache(a::T) where T
  hash = Dict{UInt,Any}()
  array_cache(hash,a)
end

for types T such that uses_hash(T) == Val(true), see the uses_hash function. In the later case, the type T should implement the following signature:

array_cache(hash::Dict,a::AbstractArray)

where we pass a dictionary (i.e., a hash table) in the first argument. This hash table can be used to test if the object a has already build a cache and re-use it as follows

id = objectid(a)
if haskey(hash,id)
  cache = hash[id] # Reuse cache
else
  cache = ... # Build a new cache depending on your needs
  hash[id] = cache # Register the cache in the hash table
end

In multi-threading computations, a different hash table per thread has to be used in order to avoid race conditions.

source

InplaceArrays.Arrays.uses_hash — Function.

uses_hash(::Type{T}) where T <:AbstractArray

This function is used to specify if the type T uses the hash-based mechanism to reuse caches. It should return either Val(true) or Val(false). It defaults to

uses_hash(::Type{<:AbstractArray}) = Val(false)

Once this function is defined for the type T it can also be called on instances of T.

Examples

julia> uses_hash(Matrix{Float64})
Val{false}()

julia> a = ones(2,3)
2×3 Array{Float64,2}:
 1.0  1.0  1.0
 1.0  1.0  1.0

julia> uses_hash(a)
Val{false}()

source

InplaceArrays.Arrays.testitem — Function.

testitem(a::AbstractArray)

Returns an arbitrary instance of eltype(a). The default returned value is the first entry in the array if length(a)>0 and testvalue(eltype(a)) if length(a)==0 See the testvalue function.

source

InplaceArrays.Arrays.testvalue — Function.

testvalue(::Type{T}) where T

Returns an arbitrary instance of type T. It defaults to zero(T) for non-array types and to an empty array for array types.

source