Cell-wise physical fields

Cell-wise physical fields

Definitions

InplaceArrays.CellFields.CellFieldLike — Type.

const CellFieldLike = CellValue{V} where V<:FieldLike{D,T,N} where {D,T,N}

source

InplaceArrays.CellFields.CellField — Type.

const CellField = CellFieldLike{D,T,1} where {D,T}

source

InplaceArrays.CellFields.CellBasis — Type.

const CellBasis = CellFieldLike{D,T,2} where {D,T}

source

InplaceArrays.CellFields.CellPoints — Type.

const CellPoints = CellValue{A} where A<:AbstractVector{P} where P<:Point{D,T} where {D,T}

source

Cell-wise evaluation and gradient

InplaceArrays.Fields.evaluate — Method.

evaluate(cf::CellFieldLike,x::CellPoints) -> CellArray

source

InplaceArrays.Fields.gradient — Method.

gradient(cf::CellFieldLike) -> CellFieldLike

source

CellFieldLike objects without underlying array

In most cases, it is not necessary to iterate over the fields stored in a cell-wise field. One can implement concrete types of the CellFieldLike type that do not have an underlying array object. In this cases, its enough to overload functions:

evaluate(cf::CellFieldLike,x::CellPoints)

gradient(cf::CellFieldLike)

Lazy operation trees

InplaceArrays.CellValues.apply — Method.

apply(g,f::CellFieldLikeOrData...) -> CellFieldLike

Apply the given functor g to f....

source

Testers

InplaceArrays.CellFields.test_cell_field_like — Function.

source

InplaceArrays.CellFields.test_cell_field_like_no_array — Function.

source

InplaceArrays.CellFields.test_cell_field — Function.

source

InplaceArrays.CellFields.test_cell_basis — Function.

source

InplaceArrays.CellFields.test_cell_field_like_with_gradient — Function.

source

InplaceArrays.CellFields.test_cell_field_like_with_gradient_no_array — Function.

source

InplaceArrays.CellFields.test_cell_field_with_gradient — Function.

source

InplaceArrays.CellFields.test_cell_basis_with_gradient — Function.

source