Gridap.Fields

Gridap.Fields

InplaceArrays.Fields — Module.

This module provides:

An interface for physical fields and basis of physical fields.

Helpers functions to work with fields and arrays of fields.

Helpers functions to create lazy operation trees from fields and arrays of fields

The exported names are:

Field

Point

apply_kernel_to_field

apply_to_field

attachmap

compose

evaluate

evaluate!

evaluate_field

evaluate_field!

evaluate_fields

evaluate_fields!

field_array_cache

field_cache

field_caches

field_gradient

field_gradients

field_return_type

field_return_types

gradient

integrate

lincomb

test_array_of_fields

test_field

varinner

∇

source

Interface

InplaceArrays.Fields.Field — Type.

abstract type Field <: Kernel

Abstract type representing physical fields, bases of fields, and other related objects.

The following functions need to be overloaded:

evaluate_field!(cache,f,x)

field_cache(f,x)

The following functions can be also provided optionally

field_gradient(f)

field_return_type(f,x)

The interface can be tested with

test_field

Most of the functionality implemented in terms of this interface relies in duck typing (this is why all functions in the interface have the word "field"). Thus, it is not strictly needed to work with types that inherit from Field. This is specially useful in order to accommodate existing types into this framework without the need to implement a wrapper type that inherits from Field. For instance, a default implementation is available for numbers, which behave like "constant" fields. However, we recommend that new types inherit from Field.

source

InplaceArrays.Fields.Point — Type.

const Point{D,T} = VectorValue{D,T}

Type representing a point of D dimensions with coordinates of type T

source

InplaceArrays.Fields.evaluate_field! — Method.

evaluate_field!(cache, f, x)

Returns an array, the length of the first axis is length(x).

source

InplaceArrays.Fields.field_cache — Method.

field_cache(f, x)

Returns the cache object needed to evaluate field f at the vector of points x.

source

InplaceArrays.Fields.field_gradient — Method.

field_gradient(f)

Returns another field that represents the gradient of the given one

source

InplaceArrays.Fields.field_return_type — Method.

field_return_type(f, x)

Computes the type obtained when evaluating field f at point x.

source

InplaceArrays.Fields.test_field — Function.

test_field(
  f,
  x::AbstractVector{<:Point},
  v::AbstractArray,cmp=(==);
  grad=nothing)

Function used to test the field interface.

source

InplaceArrays.Fields.∇ — Method.

gradient(f::Field)

Like field_gradient(f) but only for types <:Field.

The following fancy alias for the gradient function is also defined and exported.

const ∇ = gradient

source

Helper functions using fields

InplaceArrays.Fields.evaluate_field — Method.

source

InplaceArrays.Fields.evaluate — Method.

evaluate(f::Field,x)

Equivalent to

cache = field_cache(f,x)
evaluate!(cache,f,x)

source

InplaceArrays.Fields.evaluate! — Method.

source

InplaceArrays.Fields.gradient — Method.

gradient(f::Field)

Like field_gradient(f) but only for types <:Field.

The following fancy alias for the gradient function is also defined and exported.

const ∇ = gradient

source

InplaceArrays.Fields.field_return_types — Method.

source

InplaceArrays.Fields.field_caches — Method.

source

InplaceArrays.Fields.evaluate_fields — Method.

source

InplaceArrays.Fields.evaluate_fields! — Method.

source

InplaceArrays.Fields.field_gradients — Method.

source

InplaceArrays.Fields.gradient_all — Method.

source

InplaceArrays.Fields.evaluate_all — Method.

source

Applying kernels to fields

InplaceArrays.Fields.apply_kernel_to_field — Method.

apply_kernel_to_field(k::Kernel,f...) -> Field

Returns a field obtained by applying the kernel k to the values of the fields in f. That is, the returned field evaluated at a point x provides the value obtained by applying kernel k to the values of the fields f at point x. Formally, the resulting field at a point x is defined as

fx = [ evaluate(fi,x) for fi in f]
apply_kernel(k,fx...)

if any of the inputs in f is a number or an array of numbers instead of a field it will be treated as a "constant field". That is a filed that evaluated at any point x returns always the underlying number or array.

In order to be able to call the gradient function of the resulting field, one needs to define the gradient operator associated with the underlying kernel. This is done by adding a new method gradient(k::Kernel,f::Field...) for each kernel type.

source

InplaceArrays.Fields.apply_kernel_gradient — Method.

apply_kernel_gradient(k,f...)

Returns a field representing the gradient of the field obtained with

apply_kernel_to_field(k,f...)

source

Working with arrays of fields

InplaceArrays.Fields.evaluate — Method.

evaluate(a::AbstractArray{<:Field},x::AbstractArray)

Evaluates the fields in the array a at the locations provided in the array x (which can be an array of points or an array of vectors of points).

The result is numerically equivalent to

map(evaluate,a,x)

source

InplaceArrays.Fields.gradient — Method.

gradient(a::AbstractArray)

Returns an array containing the gradients of the fields in the array a. Numerically equivalent to

map(gradient,a)

source

InplaceArrays.Fields.apply_to_field — Method.

apply_to_field(k::Kernel,f::AbstractArray...)

Returns an array of fields numerically equivalent to

map( (x...) -> apply_kernel_to_field(k,x...), f )

source

InplaceArrays.Fields.kernel_evaluate — Method.

source

InplaceArrays.Fields.apply_gradient — Method.

source

InplaceArrays.Fields.field_array_cache — Method.

field_array_cache(a::AbstractArray,x::AbstractArray) -> Tuple

Returns the caches needed to perform the following iteration

ca, cfi, cx = field_array_cache(a,x)

for i in length(a)
  fi = getindex!(ca,a,i)
  xi = getindex!(cx,x,i)
  fxi = evaluate!(cfi,fi,xi)
end

source

InplaceArrays.Fields.test_array_of_fields — Function.

function test_array_of_fields(
  a::AbstractArray{<:Field},
  x::AbstractArray,
  v::AbstractArray,
  cmp::Function=(==);
  grad = nothing)

Function to test an array of fields.

source

Field operations

InplaceArrays.Fields.compose — Method.

compose(g::Function,f...)

Returns a new field obtained by composition of function g and the fields f. The value of the resulting field at point x is numerically equivalent to

fx = [ evaluate(fi,x) for fi in f]
apply_kernel(bcast(g), fx...)

The gradient of the resulting field evaluated at point x is equivalent to

fx = [ evaluate(fi,x) for fi in f]
apply_kernel(bcast(gradient(g)), fx...)

Note that it is needed to overload gradient(::typeof(g)) for the given function g in order to be able to compute the gradient.

As in function apply_kernel_to_field if any of the inputs in f is a number or an array instead of a field it will be treated as a "constant field".

source

InplaceArrays.Fields.compose — Method.

compose(g::Function,f::AbstractArray...)

Returns an array of fields numerically equivalent to

map( (x...)->compose(g,x...), f...)

source

InplaceArrays.Fields.lincomb — Method.

lincomb(a::Basis,b::AbstractVector)

Returns a field f with valuetype(f) <: Number obtained by the "linear combination" of the value of the basis a and the vector b. That is, the value of the resulting field f at a point x is defined as

k = contract(outer)
ax = evaluate(a,x)
apply_kernel(k,ax,b)

On the other hand, the gradient of the resulting field is defined as

k = contract(outer)
∇ax = evaluate(gradient(a),x)
apply_kernel(k,∇ax,b)

source

InplaceArrays.Fields.lincomb — Method.

lincomb(a::AbstractArray{<:Field},b::AbstractArray)

Returns an array of field numerically equivalent to

map(lincomb,a,b)

source

InplaceArrays.Fields.varinner — Method.

source

InplaceArrays.Fields.varinner — Method.

source

InplaceArrays.Fields.attachmap — Method.

source

InplaceArrays.Fields.attachmap — Method.

source

InplaceArrays.Fields.integrate — Method.

source

InplaceArrays.Fields.integrate — Method.

source