The functor interface

The Functor interface

Often, it is needed to implement functions that need some scratch data (e.g., pre-allocating the output). The question is, where to store this data? There are three main answers to this question: 1) store the data in the function object as part of its state, 2) allocate the scratch data each time the operation is performed, and 3) the user allocates and passes the scratch data when needed. Clearly, 1) it is not save if several calls to the operation are using the same scratch data (e.g., multi-threading). 2) is save, but it can be inefficient if the operation is performed at low granularity. 3) is both save and efficient, but requires some extra work by the user.

In Gridap, we adopt the 3rd option. In order to unify the interfaces of functions using this approach, we introduce the Functor interface. Any type is referred to as a Functor if it implements the following interface. We rely in duck typing here. There is not an abstract type representing a functor.

InplaceArrays.Functors.evaluate_functor! — Function.

y = evaluate_functor!(cache,f,x...)

Evaluates the functor f at the arguments x... using the scratch data provided in the given cache object. The cache object is built with the functor_cache function using arguments of the same type as in x... In general, the returned value y can share some part of its state with the cache object. If the result of two or more invocations of this function need to be accessed simultaneously (e.g., in multi-threading), create and use various cache objects (e.g., one cache per thread).

source

InplaceArrays.Functors.functor_cache — Function.

cache = functor_cache(f,x...)

Returns the cache needed to evaluate functor f with arguments of the same type as the objects in x....

source

Default implementations

We provide some default (obvious) implementations of this interface so that Function, Number, and AbstractArray objects behave like functors.

Examples

Calling the + function via the functor interface.

julia> cache = functor_cache(+,0,0)

julia> evaluate_functor!(cache,+,1,2)
3

julia> evaluate_functor!(cache,+,-1,10)
9

Number and AbstractArray objects behave like "constant" functors.

julia> a = 2.0
2.0

julia> cache = functor_cache(a,0)

julia> evaluate_functor!(cache,a,1)
2.0

julia> evaluate_functor!(cache,a,2)
2.0

julia> evaluate_functor!(cache,a,3)
2.0

Evaluating a functor without cache

InplaceArrays.Functors.evaluate_functor — Function.

evaluate_functor(f,x...)

Evaluate the fuctor f at the arguments x... by creating a temporary cache internally. This functions is equivalent to

cache = functor_cache(f,x...)
evaluate_functor!(cache,f,x...)

source

Working with several functors at once

InplaceArrays.Functors.evaluate_functors! — Function.

evaluate_functors!(caches::Tuple,fs::Tuple,x...) -> Tuple

Evaluates the functors in the tuple fs at the arguments x... by using the corresponding cache objects in the tuple caches. The result is also a tuple containing the result for each functor in fs.

source

InplaceArrays.Functors.functor_caches — Function.

functor_caches(fs::Tuple,x...) -> Tuple

Returns a tuple with the cache corresponding to each functor in fs for the arguments x....

source

Broadcasting

InplaceArrays.Functors.bcast — Function.

bcast(f::Function)

Returns a functor object that represents the "boradcasted" version of the given function f.

Examples

julia> op = bcast(*)
InplaceArrays.Functors.BCasted{typeof(*)}(*)

julia> x = ones(2,3)
2×3 Array{Float64,2}:
 1.0  1.0  1.0
 1.0  1.0  1.0

julia> y = 2
2

julia> evaluate_functor(op,x,y)
2×3 CachedArray{Float64,2,Array{Float64,2}}:
 2.0  2.0  2.0
 2.0  2.0  2.0

source

Composition

InplaceArrays.Functors.compose_functors — Function.

c = compose_functors(g,fs...)

Returns an object c representing the "composition" of functor g with several functors fs. The resulting object c is such that

evaluate_functor(c,x...)

is equivalent to

fxs = evaluate_functors(fs,x...)
evaluate_functor(g,fxs...)

source