Gridap.Helpers

Gridap.Helpers

InplaceArrays.Helpers — Module.

This module provides a set of helper macros and helper functions

The exported macros are:

@abstractmethod

@notimplemented

@notimplementedif

tfill

@unreachable

source

InplaceArrays.Helpers.tfill — Method.

tfill(v, ::Val{D}) where D

Returns a tuple of length D that contains D times the object v. In contrast to tuple(fill(v,D)...) which returns the same result, this function is type-stable.

source

InplaceArrays.Helpers.@abstractmethod — Macro.

@abstractmethod

Macro used in generic functions that must be overloaded by derived types.

source

InplaceArrays.Helpers.@notimplemented — Macro.

@notimplemented
@notimplemented "Error message"

Macro used to raise an error, when something is not implemented.

source

InplaceArrays.Helpers.@notimplementedif — Macro.

@notimplementedif condition
@notimplementedif condition "Error message"

Macro used to raise an error if the condition is true

source

InplaceArrays.Helpers.@unreachable — Macro.

@unreachable
@unreachable "Error message"

Macro used to make sure that a line of code is never reached.

source