Gridap.Polynomials

Gridap.Polynomials

InplaceArrays.Polynomials — Module.

This module provides a collection of multivariate polynomial bases.

The exported names are:

MonomialBasis

QCurlGradMonomialBasis

QGradMonomialBasis

change_basis

source

InplaceArrays.Polynomials.MonomialBasis — Type.

struct MonomialBasis{...} <: Field

Type representing a basis of multivariate scalar-valued, vector-valued, or tensor-valued, iso- or aniso-tropic monomials. The type parameters and fields of this struct are not public, they are private implementation details. This type fully implements the Field interface, with up to second order derivatives.

source

InplaceArrays.Polynomials.MonomialBasis — Method.

MonomialBasis{D}(::Type{T}, order::Int [, filter::Function]) where {D,T}

Returns an instance of MonomialBasis representing a multivariate polynomial basis in D dimensions, of polynomial degree order, whose value is represented by the type T. The type T is typically <:Number, e.g., Float64 for scalar-valued functions and VectorValue{D,Float64} for vector-valued ones.

Filter function

The filter function is used to select which terms of the tensor product space of order order in D dimensions are to be used. If the filter is not provided, the full tensor-product space is used by default leading to a multivariate polynomial space of type Q. The signature of the filter function is

(e,order) -> Bool

where e is a tuple of D integers containing the exponents of a multivariate monomial. The following filters are used to select well known polynomial spaces

Q space: (e,order) -> true

P space: (e,order) -> sum(e) <= order

"Serendipity" space: (e,order) -> sum( ( i for i in e if i>1 ) ) <= order

source

InplaceArrays.Polynomials.MonomialBasis — Method.

MonomialBasis{D}(::Type{T}, orders::Tuple [, filter::Function]) where {D,T}

This version of the constructor allows to pass a tuple orders containing the polynomial order to be used in each of the D dimensions in order to construct and anisotropic tensor-product space.

source

InplaceArrays.Polynomials.QCurlGradMonomialBasis — Type.

struct QCurlGradMonomialBasis{...} <: Field

This type implements a multivariate vector-valued polynomial basis spanning the space needed for Raviart-Thomas reference elements on n-cubes. The type parameters and fields of this struct are not public. This type fully implements the Field interface, with up to first order derivatives.

source

InplaceArrays.Polynomials.QCurlGradMonomialBasis — Method.

QCurlGradMonomialBasis{D}(::Type{T},order::Int) where {D,T}

Returns a QCurlGradMonomialBasis object. D is the dimension of the coordinate space and T is the type of the components in the vector-value.

source

InplaceArrays.Polynomials.QGradMonomialBasis — Type.

struct QGradMonomialBasis{...} <: Field

This type implements a multivariate vector-valued polynomial basis spanning the space needed for Nedelec reference elements on n-cubes. The type parameters and fields of this struct are not public. This type fully implements the Field interface, with up to first order derivatives.

source

InplaceArrays.Polynomials.QGradMonomialBasis — Method.

QGradMonomialBasis{D}(::Type{T},order::Int) where {D,T}

Returns a QGradMonomialBasis object. D is the dimension of the coordinate space and T is the type of the components in the vector-value.

source

InplaceArrays.Polynomials.change_basis — Method.

source