Gridap.ReferenceFEs

Gridap.ReferenceFEs

InplaceArrays.ReferenceFEs — Module.

The exported names are

Dof

ExtrusionPolytope

GenericRefFE

HEX

HEX_AXIS

LagrangianDofBasis

LagrangianRefFE

PYRAMID

Polytope

QUAD

ReferenceFE

SEGMENT

SerendipityRefFE

TET

TET_AXIS

TRI

VERTEX

WEDGE

compute_shapefuns

dof_cache

dof_return_type

edge_tangents

evaluate_dof

evaluate_dof!

evaluate_dof_array

facet_normals

facet_orientations

num_dims

num_edges

num_faces

num_facets

num_vertices

polytope_dimrange

polytope_facedims

polytope_faces

polytope_offset

polytope_offsets

polytope_vtkid

polytope_vtknodes

reffe_dof_permutations

reffe_dofs

reffe_face_dofids

reffe_polytope

reffe_prebasis

reffe_shapefuns

test_dof

test_polytope

test_reference_fe

vertex_coordinates

vertex_permutations

source

Polytopes

Interface

InplaceArrays.ReferenceFEs.Polytope — Type.

abstract type Polytope{D}

Abstract type representing a polytope (i.e., a polyhedron in arbitrary dimensions). D is the environment dimension (typically, 0, 1, 2, or 3). This type parameter is needed since there are functions in the Polytope interface that return containers with Point{D} objects. We adopt the usual nomenclature for polytope-related objects. All objects in a polytope (from vertices to the polytope itself) are called n-faces or simply faces. The notation n-faces is used only when it is needed to refer to the object dimension n. Otherwise we simply use face. In addition, we say

vertex (pl. vertices): for 0-faces

edge: for 1-faces

facet: for (D-1)-faces

The Polytope interface is defined by overloading the following functions

polytope_faces(p)

polytope_dimrange(p)

Polytope{N}(p,faceid) where N

vertex_coordinates(p)

(==)(a::Polytope{D},b::Polytope{D}) where D

And optionally these ones:

edge_tangents(p)

facet_normals(p)

facet_orientations(p)

vertex_permutations(p)

polytope_vtkid(p)

polytope_vtknodes(p)

The interface can be tested with the function

test_polytope

source

InplaceArrays.ReferenceFEs.polytope_faces — Method.

polytope_faces(p) -> Vector{Vector{Int}}

Given a polytope p the function returns a vector of vectors defining the incidence relation of the faces in the polytope.

Each face in the polytope receives a unique integer id. The id 1 is assigned to the first 0-face. Consecutive increasing ids are assigned to the other 0-faces, then to 1-faces, and so on. The polytope itself receives the largest id which coincides with num_faces(p). For a face id iface, polytope_faces(p)[iface] is a vector of face ids, corresponding to the faces that are incident with the face labeled with iface. That is, faces that are either on its boundary or the face itself. In this vector of incident face ids, faces are ordered by dimension, starting with 0-faces. Within each dimension, the labels are ordered in a consistent way with the polyope object for the face iface itself.

Examples

using InplaceArrays.ReferenceFEs

faces = polytope_faces(SEGMENT)
println(faces)

# output
Array{Int64,1}[[1], [2], [1, 2, 3]]

The constant SEGMENT is bound to a predefined instance of polytope that represents a segment. The face labels associated with a segment are [1,2,3], being 1 and 2 for the vertices and 3 for the segment itself. In this case, this function would return the vector of vectors [[1],[2],[1,2,3]] meaning that vertex 1 is incident with vertex 1 (idem for vertex 2), and that the segment (id 3) is incident with the vertices 1 and 2 and the segment itself.

source

InplaceArrays.ReferenceFEs.polytope_dimrange — Method.

polytope_dimrange(p) -> Vector{UnitRange{Int}}

Given a polytope p it returns a vector of ranges. The entry d+1 in this vector contains the range of face ids for the faces of dimension d.

Examples

using InplaceArrays.ReferenceFEs

ranges = polytope_dimrange(SEGMENT)
println(ranges)

# output
UnitRange{Int64}[1:2, 3:3]

Face ids for the vertices in the segment range from 1 to 2 (2 vertices), the face ids for edges in the segment range from 3 to 3 (only one edge with id 3).

source

InplaceArrays.ReferenceFEs.Polytope — Method.

Polytope{N}(p,faceid) where N

Returns a Polytope{N} object representing the "reference" polytope of the N-face with id faceid. The value faceid refers to the numeration restricted to the dimension N (it starts with 1 for the first N-face).

source

InplaceArrays.ReferenceFEs.vertex_coordinates — Method.

vertex_coordinates(p) -> Vector{Point{D,Float64}}

Given a polytope p return a vector of points representing containing the coordinates of the vertices.

source

Base.:== — Method.

(==)(a::Polytope{D},b::Polytope{D}) where D

Returns true if the polytopes a and b are equivalent. Otherwise, it returns false. Note that the operator == returns false by default for polytopes of different dimensions. Thus, this function has to be overloaded only for the case of polytopes a and b of same dimension.

source

InplaceArrays.ReferenceFEs.edge_tangents — Method.

edge_tangents(p) -> Vector{VectorValue{D,Float64}}

Given a polytope p, returns a vector of VectorValue objects representing the unit tangent vectors to the polytope edges.

source

InplaceArrays.ReferenceFEs.facet_normals — Method.

facet_normals(p) -> Vector{VectorValue{D,Float64}}

Given a polytope p, returns a vector of VectorValue objects representing the unit outward normal vectors to the polytope facets.

source

InplaceArrays.ReferenceFEs.facet_orientations — Method.

facet_orientations(p) -> Vector{Int}

Given a polytope p returns a vector of integers of length num_facets(p). Facets, whose vertices are ordered consistently with the outwards normal vector, receive value 1 in this vector. Otherwise, facets receive value -1.

source

InplaceArrays.ReferenceFEs.vertex_permutations — Method.

vertex_permutations(p) -> Vector{Vector{Int}}

Given a polytope p, returns a vector of vectors containing all admissible permutations of the polytope vertices. An admissible permutation is one such that, if the vertices of the polytope are re-labeled according to this permutation, the resulting polytope preserves the shape of the original one.

Examples

using InplaceArrays.ReferenceFEs

perms = vertex_permutations(SEGMENT)
println(perms)

# output
Array{Int64,1}[[1, 2], [2, 1]]

The first admissible permutation for a segment is [1,2],i.e., the identity. The second one is [2,1], i.e., the first vertex is relabeled as 2 and the second vertex is relabeled as 1.

source

InplaceArrays.ReferenceFEs.polytope_vtkid — Method.

polytope_vtkid(p) -> Int

Given a polytope p, returns an integer with its vtk identifier. Overloading of this function is needed only in order to visualize the underlying polytope with Paraview.

source

InplaceArrays.ReferenceFEs.polytope_vtknodes — Method.

polytope_vtknodes(p) -> Vector{Int}

Given a polytope p, returns a vector of integers representing a permutation of the polytope vertices required to relabel the vertices according the criterion adopted in Paraview. Overloading of this function is needed only in order to visualize the underlying polytope with Paraview.

source

InplaceArrays.ReferenceFEs.test_polytope — Method.

test_polytope(p::Polytope{D}; optional::Bool=false) where D

Function that stresses out the functions in the Polytope interface. It tests whether the function in the polytope interface are defined for the given object, and whether they return objects of the expected type. With optional=false (the default), only the mandatory functions are checked. With optional=true, the optional functions are also tested except polytope_vtkid and polytope_vtknodes.

source

InplaceArrays.ReferenceFEs.num_dims — Method.

num_dims(::Type{<:Polytope{D}}) where D
num_dims(p::Polytope{D}) where D

Returns D.

source

source

InplaceArrays.ReferenceFEs.num_faces — Method.

num_faces(p::Polytope)

Returns the total number of faces in polytope p (from vertices to the polytope itself).

source

InplaceArrays.ReferenceFEs.num_faces — Method.

num_faces(p::Polytope,dim::Integer)

Returns the number of faces of dimension dim in polytope p.

source

InplaceArrays.ReferenceFEs.num_facets — Method.

num_facets(p::Polytope)

Returns the number of facets in the polytope p.

source

InplaceArrays.ReferenceFEs.num_edges — Method.

num_edges(p::Polytope)

Returns the number of edges in the polytope p.

source

InplaceArrays.ReferenceFEs.num_vertices — Method.

num_vertices(p::Polytope)

Returns the number of vertices in the polytope p.

source

InplaceArrays.ReferenceFEs.polytope_facedims — Method.

polytope_facedims(p) -> Vector{Int}

Given a polytope p, returns a vector indicating the dimension of each face in the polytope

Examples

using InplaceArrays.ReferenceFEs

dims = polytope_facedims(SEGMENT)
println(dims)

# output
[0, 0, 1]

The first two faces in the segment (the two vertices) have dimension 0 and the third face (the segment itself) has dimension 1

source

InplaceArrays.ReferenceFEs.polytope_offsets — Method.

polytope_offsets(p) -> Vector{Int}

Given a polytope p, it returns a vector of integers. The position in the d+1 entry in this vector is the offset that transforms a face id in the global numeration in the polytope to the numeration restricted to faces to dimension d.

Examples

using InplaceArrays.ReferenceFEs

offsets = polytope_offsets(SEGMENT)
println(offsets)

# output
[0, 2]

source

InplaceArrays.ReferenceFEs.polytope_offset — Method.

polytope_offset(p,d)

Equivalent to polytope_offsets(p)[d+1].

source

InplaceArrays.ReferenceFEs.polytope_faces — Method.

polytope_faces(p,dimfrom,dimto) -> Vector{Vector{Int}}

For dimfrom >= dimto returns a vector that for each face of dimension dimfrom stores a vector of the ids of faces of dimension dimto on its boundary.

For dimfrom < dimto returns a vector that for each face of dimfrom stores a vector of the face ids of faces of dimension dimto that touch it.

The numerations used in this funcitons are the ones restricted to each dimension.

using InplaceArrays.ReferenceFEs

edge_to_vertices = polytope_faces(QUAD,1,0)
println(edge_to_vertices)

vertex_to_edges_around = polytope_faces(QUAD,0,1)
println(vertex_to_edges_around)

# output
Array{Int64,1}[[1, 2], [3, 4], [1, 3], [2, 4]]
Array{Int64,1}[[1, 3], [1, 4], [2, 3], [2, 4]]

source

Extrusion polytopes

InplaceArrays.ReferenceFEs.ExtrusionPolytope — Type.

struct ExtrusionPolytope{D} <: Polytope{D}
  extrusion::Point{D,Int}
  # + private fields
end

Concrete type for polytopes that can be represented with an "extrusion" tuple. The underlying extrusion is available in the field extrusion. Instances of this type can be obtained with the constructors

Polytope(extrusion::Int...)

ExtrusionPolytope(extrusion::Int...)

source

InplaceArrays.ReferenceFEs.ExtrusionPolytope — Method.

ExtrusionPolytope(extrusion::Int...)

Generates an ExtrusionPolytope from the tuple extrusion. The values in extrusion are either equal to the constant HEX_AXIS or the constant TET_AXIS.

Examples

Creating a quadrilateral, a triangle, and a wedge

using InplaceArrays.ReferenceFEs

quad = ExtrusionPolytope(HEX_AXIS,HEX_AXIS)

tri = ExtrusionPolytope(TET_AXIS,TET_AXIS)

wedge = ExtrusionPolytope(TET_AXIS,TET_AXIS,HEX_AXIS)

println(quad == QUAD)
println(tri == TRI)
println(wedge == WEDGE)

# output
true
true
true

source

InplaceArrays.ReferenceFEs.Polytope — Method.

Polytope(extrusion::Int...)

Equivalent to ExtrusionPolytope(extrusion...)

source

InplaceArrays.ReferenceFEs.HEX_AXIS — Constant.

Constant to be used in order to indicate a hex-like extrusion axis.

source

InplaceArrays.ReferenceFEs.TET_AXIS — Constant.

Constant to be used in order to indicate a tet-like extrusion axis.

source

Pre-defined polytope instances

InplaceArrays.ReferenceFEs.VERTEX — Constant.

Constant bound to a polytope instance representing a vertex.

source

InplaceArrays.ReferenceFEs.SEGMENT — Constant.

const SEGMENT = Polytope(HEX_AXIS)

source

InplaceArrays.ReferenceFEs.TRI — Constant.

const TRI = Polytope(TET_AXIS,TET_AXIS)

source

InplaceArrays.ReferenceFEs.QUAD — Constant.

const QUAD = Polytope(HEX_AXIS,HEX_AXIS)

source

InplaceArrays.ReferenceFEs.TET — Constant.

const TET = Polytope(TET_AXIS,TET_AXIS,TET_AXIS)

source

InplaceArrays.ReferenceFEs.HEX — Constant.

const HEX = Polytope(HEX_AXIS,HEX_AXIS,HEX_AXIS)

source

InplaceArrays.ReferenceFEs.WEDGE — Constant.

const WEDGE = Polytope(TET_AXIS,TET_AXIS,HEX_AXIS)

source

InplaceArrays.ReferenceFEs.PYRAMID — Constant.

const PYRAMID = Polytope(HEX_AXIS,HEX_AXIS,TET_AXIS)

source

Degrees of freedom

Interface

InplaceArrays.ReferenceFEs.Dof — Type.

abstract type Dof <: Kernel

Abstract type representing a degree of freedom (DOF), a basis of DOFs, and related objects. These different cases are distinguished by the return type obtained when evaluating the Dof object on a Field object. See function evaluate_dof! for more details.

The following functions needs to be overloaded

dof_cache

evaluate_dof!

The following functions can be overloaded optionally

dof_return_type

The interface is tested with

test_dof

In most of the cases it is not strictly needed that types that implement this interface inherit from Dof. However, we recommend to inherit from Dof, when possible.

source

InplaceArrays.ReferenceFEs.evaluate_dof! — Method.

evaluate_dof!(cache,dof,field)

Evaluates the dof dof with the field field. It can return either an scalar value or an array of scalar values depending the case. The cache object is computed with function dof_cache.

When a mathematical dof is evaluated on a physical field, a scalar number is returned. If either the Dof object is a basis of DOFs, or the Field object is a basis of fields, or both objects are bases, then the returned object is an array of scalar numbers. The first dimensions in the resulting array are for the Dof object and the last ones for the Field object. E.g, a basis of nd DOFs evaluated at physical field returns a vector of nd entries. A basis of nd DOFs evaluated at a basis of nf fields returns a matrix of size (nd,nf).

source

InplaceArrays.ReferenceFEs.dof_cache — Method.

dof_cache(dof,field)

Returns the cache needed to call evaluate_dof!(cache,dof,field)

source

InplaceArrays.ReferenceFEs.dof_return_type — Method.

dof_return_type(dof,field)

Returns the type for the value obtained with evaluating dof with field.

It defaults to

typeof(evaluate_dof(dof,field))

source

InplaceArrays.ReferenceFEs.test_dof — Method.

test_dof(dof,field,v,comp::Function=(==))

Test that the Dof interface is properly implemented for object dof. It also checks if the object dof when evaluated at the field field returns the same value as v. Comparison is made with the comp function.

source

InplaceArrays.ReferenceFEs.evaluate_dof — Method.

evaluate_dof(dof,field)

Equivalent to

cache = dof_cache(dof,field)
evaluate_dof!(cache,dof,field)

source

InplaceArrays.Fields.evaluate — Method.

evaluate(dof::Dof,field)

Equivalent to evaluate_dof(dof,field).

source

Working with arrays of DOFs

InplaceArrays.ReferenceFEs.evaluate_dof_array — Method.

evaluate_dof_array(dof::AbstractArray,field::AbstractArray)

Evaluates the Dof objects in the array dof at the Field objects at the array field element by element.

The result is numerically equivalent to

map(evaluate_dof, dof, field)

but it is described with a more memory-friendly lazy type.

source

InplaceArrays.Fields.evaluate — Method.

evaluate(dof::AbstractArray{<:Dof},field::AbstractArray)

Equivalent to evaluate_dof_array(dof,field)

source

Lagrangian dof bases

InplaceArrays.ReferenceFEs.LagrangianDofBasis — Type.

struct LagrangianDofBasis{P,V} <: Dof
  nodes::Vector{P}
  dof_to_node::Vector{Int}
  dof_to_comp::Vector{Int}
  node_and_comp_to_dof::Vector{V}
end

Type that implements a Lagrangian dof basis.

Fields:

nodes::Vector{P} vector of points (P<:Point) storing the nodal coordinates

node_and_comp_to_dof::Vector{V} vector such that node_and_comp_to_dof[node][comp] returns the dof associated with node node and the component comp in the type V.

dof_to_node::Vector{Int} vector of integers such that dof_to_node[dof] returns the node id associated with dof id dof.

dof_to_comp::Vector{Int} vector of integers such that dof_to_comp[dof] returns the component id associated with dof id dof.

source

InplaceArrays.ReferenceFEs.LagrangianDofBasis — Method.

LagrangianDofBasis(::Type{T},nodes::Vector{<:Point}) where T

Creates a LagrangianDofBasis for fields of value type T associated with the vector of nodal coordinates nodes.

source

Reference Finite Elements

Interface

InplaceArrays.ReferenceFEs.ReferenceFE — Type.

source

InplaceArrays.ReferenceFEs.num_dofs — Method.

source

InplaceArrays.ReferenceFEs.reffe_polytope — Method.

source

InplaceArrays.ReferenceFEs.reffe_prebasis — Method.

source

InplaceArrays.ReferenceFEs.reffe_dofs — Method.

source

InplaceArrays.ReferenceFEs.reffe_face_dofids — Method.

source

InplaceArrays.ReferenceFEs.ReferenceFE — Method.

source

InplaceArrays.ReferenceFEs.reffe_dof_permutations — Method.

source

InplaceArrays.ReferenceFEs.INVALID_PERM — Constant.

source

InplaceArrays.ReferenceFEs.reffe_shapefuns — Method.

source

InplaceArrays.ReferenceFEs.compute_shapefuns — Method.

source

InplaceArrays.ReferenceFEs.num_dims — Method.

num_dims(::Type{<:Polytope{D}}) where D
num_dims(p::Polytope{D}) where D

Returns D.

source

source

InplaceArrays.ReferenceFEs.test_reference_fe — Method.

source

Generic reference elements

InplaceArrays.ReferenceFEs.GenericRefFE — Type.

struct GenericRefFE{D} <: ReferenceFE{D}
  ndofs::Int
  polytope::Polytope{D}
  prebasis::Field
  dofs::Dof
  facedofids::Vector{Vector{Int}}
  shapefuns::Field
  reffaces::Tuple
end

This type is a materialization of the ReferenceFE interface. That is, it is a struct that stores the values of all abstract methods in the ReferenceFE interface. This type is useful to build reference FEs from the underlying ingredients without the need to create a new type.

Note that this struct is type unstable deliberately in order to simplify the type signature and speed up compilation times. Don't use it in computationally expensive functions, instead extract the required fields and pass them to the computationally expensive function.

source

InplaceArrays.ReferenceFEs.GenericRefFE — Method.

source

Lagrangian reference elements

InplaceArrays.ReferenceFEs.LagrangianRefFE — Type.

struct LagrangianRefFE{D} <: ReferenceFE{D}
  data::GenericRefFE{D}
  facenodeids::Vector{Vector{Int}}
  nodeperms::Vector{Vector{Int}},
end

For this type

reffe_dofs(reffe) returns a LagrangianDofBasis

reffe_prebasis(reffe) returns a MonomialBasis

ReferenceFE{N}(reffe,faceid) where N returns a LagrangianRefFE{N}

source

InplaceArrays.ReferenceFEs.LagrangianRefFE — Method.

LagrangianRefFE(::Type{T},p::Polytope,orders) where T
LagrangianRefFE(::Type{T},p::Polytope,order::Int) where T

source

InplaceArrays.Polynomials.MonomialBasis — Method.

source

InplaceArrays.ReferenceFEs.LagrangianDofBasis — Method.

source

InplaceArrays.ReferenceFEs.compute_monomial_basis — Method.

source

InplaceArrays.ReferenceFEs.compute_interior_nodes — Method.

source

InplaceArrays.ReferenceFEs.compute_face_orders — Method.

source

InplaceArrays.ReferenceFEs.compute_nodes — Method.

source

InplaceArrays.ReferenceFEs.compute_node_permutations — Method.

source

InplaceArrays.ReferenceFEs.compute_lagrangian_reffaces — Method.

source

Serendipity reference elements

InplaceArrays.ReferenceFEs.SerendipityRefFE — Function.

source

InplaceArrays.ReferenceFEs.is_serendipity_compatible — Method.

source