Gridap.Arrays

Gridap.Arrays

InplaceArrays.Arrays — Module.

This module provides:

An extension of the AbstractArray interface in order to properly deal with mutable caches.

A mechanism to generate lazy arrays resulting from operations between arrays.

A collection of concrete implementations of AbstractArray.

The exported names in this module are:

CachedArray

CachedMatrix

CachedVector

apply

apply_all

apply_kernel

apply_kernel!

apply_kernels!

array_cache

array_caches

bcast

contract

elem

getindex!

getitems!

kernel_cache

kernel_caches

kernel_return_type

kernel_return_types

setsize!

test_array

test_kernel

testitem

testitems

uses_hash

source

Extended AbstractArray interface

New methods added that can be overload by new types:

getindex!(cache,a::AbstractArray,i...)

array_cache(a::AbstractArray)

uses_hash(::Type{<:AbstractArray})

testitem(a::AbstractArray)

The interface can be tested with the following function

test_array

InplaceArrays.Arrays.getindex! — Method.

getindex!(cache::Any, a::AbstractArray, i::Vararg{Any,N} where N) -> Any

Returns the item of the array a associated with index i by (possibly) using the scratch data passed in the cache object.

It defaults to

getindex!(cache,a::AbstractArray,i...) = a[i...]

source

InplaceArrays.Arrays.getitems! — Function.

getitems!(cf, a, i)

source

InplaceArrays.Arrays.array_cache — Method.

array_cache(a::AbstractArray) -> Any

Returns a cache object to be used in the getindex! function. It defaults to

array_cache(a::T) where T = nothing

for types T such that uses_hash(T) == Val(false), and

function array_cache(a::T) where T
  hash = Dict{UInt,Any}()
  array_cache(hash,a)
end

for types T such that uses_hash(T) == Val(true), see the uses_hash function. In the later case, the type T should implement the following signature:

array_cache(hash::Dict,a::AbstractArray)

where we pass a dictionary (i.e., a hash table) in the first argument. This hash table can be used to test if the object a has already build a cache and re-use it as follows

id = objectid(a)
if haskey(hash,id)
  cache = hash[id] # Reuse cache
else
  cache = ... # Build a new cache depending on your needs
  hash[id] = cache # Register the cache in the hash table
end

This mechanism is needed, e.g., to re-use intermediate results in complex lazy operation trees. In multi-threading computations, a different hash table per thread has to be used in order to avoid race conditions.

source

InplaceArrays.Arrays.array_caches — Function.

array_caches(a, b)

source

InplaceArrays.Arrays.uses_hash — Method.

uses_hash(::Type{<:AbstractArray})

This function is used to specify if the type T uses the hash-based mechanism to reuse caches. It should return either Val(true) or Val(false). It defaults to

uses_hash(::Type{<:AbstractArray}) = Val(false)

Once this function is defined for the type T it can also be called on instances of T.

source

InplaceArrays.Arrays.testitem — Method.

Returns an arbitrary instance of eltype(a). The default returned value is the first entry in the array if length(a)>0 and testvalue(eltype(a)) if length(a)==0 See the testvalue function.

source

InplaceArrays.Arrays.testitems — Function.

testitems(b::AbstractArray...) -> Tuple

Returns a tuple with the result of testitem applied to each of the arrays in b.

source

InplaceArrays.Arrays.test_array — Function.

test_array(
  a::AbstractArray{T,N}, b::AbstractArray{S,N},cmp=(==)) where {T,S,N}

Checks if the entries in a and b are equal using the comparison function cmp. It also stresses the new methods added to the AbstractArray interface interface.

source

Creting lazy operation trees

InplaceArrays.Arrays.apply — Method.

apply(f,a::AbstractArray...) -> AbstractArray

Applies the kernel f to all entries in the arrays in a.

The resulting array has the same entries as the one obtained with (see function apply_kernel):

map( (x...)->apply_kernel(f,x...), a...)

Examples

using InplaceArrays.Arrays
a = apply(+,[1,2,3],[4,5,6])
println(a)
# output
[5, 7, 9]

source

InplaceArrays.Arrays.apply — Method.

apply(f::AbstractArray,a::AbstractArray...) -> AbstractArray

Applies the kernels in the array of kernels f to the entries in the arrays in a.

The resulting array has the same entries as the one obtained with (see function apply_kernel):

map( apply_kernel, f, a...)

Examples

using InplaceArrays.Arrays
a = apply([+,-,mod],[1,2,3],[4,5,6])
println(a)
# output
[5, -3, 3]

source

InplaceArrays.Arrays.apply_all — Function.

apply_all(f::Tuple,a::AbstractArray...)

source

Operation kernels

The apply function provides a mechanism to construct lazy arrays obtained by applying some operations to other arrays. The operations are represented by objects (referred to as kernels). We rely in duck typing here. There is not an abstract type representing a kernel. Any type is referred to as a kernel if it implements the following interface:

apply_kernel!(cache,k,x...)

kernel_cache(k,x...)

kernel_return_type(k,x...)

The kernel interface can be tested with the test_kernel function.

We provide some default (obvious) implementations of this interface so that Function, Number, and AbstractArray objects behave like kernels.

julia> using InplaceArrays.Arrays

julia> cache = kernel_cache(+,0,0)

julia> apply_kernel!(cache,+,1,2)
3

julia> apply_kernel!(cache,+,-1,10)
9

Number and AbstractArray objects behave like "constant" kernels.

julia> using InplaceArrays.Arrays

julia> a = 2.0
2.0

julia> cache = kernel_cache(a,0)

julia> apply_kernel!(cache,a,1)
2.0

julia> apply_kernel!(cache,a,2)
2.0

julia> apply_kernel!(cache,a,3)
2.0

InplaceArrays.Arrays.apply_kernel! — Function.

apply_kernel!(cache, f, x)

applies the kernel f at the arguments x... using the scratch data provided in the given cache object. The cache object is built with the kernel_cache function using arguments of the same type as in x... In general, the returned value y can share some part of its state with the cache object. If the result of two or more invocations of this function need to be accessed simultaneously (e.g., in multi-threading), create and use various cache objects (e.g., one cache per thread).

source

InplaceArrays.Arrays.kernel_cache — Function.

kernel_cache(f, x)

Returns the cache needed to apply kernel f with arguments of the same type as the objects in x....

source

InplaceArrays.Arrays.kernel_return_type — Function.

kernel_return_type(f, x)

Returns the type of the result of calling kernel f with arguments of the types of the objects x.

source

InplaceArrays.Arrays.test_kernel — Function.

test_kernel(f, x, y)
test_kernel(f, x, y, cmp)

Function used to test if the kernel f has been implemented correctly. f is a kernel object, x is the input of the kernel, and y is the expected result. Function cmp is used to compare the computed result with the expected one. The checks are done with the @test macro.

source

Build-in kernels

InplaceArrays.Arrays.bcast — Function.

bcast(f::Function)

Returns a kernel object that represents the "boradcasted" version of the given function f.

source

InplaceArrays.Arrays.elem — Function.

elem(f::Function)

Returns a kernel that represents the element-wise version of the binary or unary operation f

source

InplaceArrays.Arrays.contract — Function.

contract(f::Function)

Like the dot product between to vectors, but using operation f instead of * between components.

Examples

using InplaceArrays.Arrays
k = contract(-)
apply_kernel(k,[1,2],[2,4]) # Equivalent to (1-2) + (2-4)
# output
-3

source

Other functions acting on kernels

InplaceArrays.Arrays.apply_kernel — Function.

apply_kernel(f,x...)

apply the fuctor f at the arguments x... by creating a temporary cache internally. This functions is equivalent to

cache = kernel_cache(f,x...)
apply_kernel!(cache,f,x...)

source

InplaceArrays.Arrays.apply_kernels! — Function.

apply_kernels!(caches::Tuple,fs::Tuple,x...) -> Tuple

Applies the kernels in the tuple fs at the arguments x... by using the corresponding cache objects in the tuple caches. The result is also a tuple containing the result for each kernel in fs.

source

InplaceArrays.Arrays.kernel_caches — Function.

kernel_caches(fs::Tuple,x...) -> Tuple

Returns a tuple with the cache corresponding to each kernel in fs for the arguments x....

source

InplaceArrays.Arrays.kernel_return_types — Function.

kernel_return_types(f, Ts)

source

Concrete array implementations

CachedArray

InplaceArrays.Arrays.CachedArray — Type.

mutable struct CachedArray{T, N, A<:AbstractArray{T,N}} <: AbstractArray{T,N}

Type providing a re-sizable array that only allocates memory when the underlying buffer needs to grow.

The size of a CachedArray is changed via the setsize! function.

A CachedArray can be build with the constructors

CachedArray(a::AbstractArray)

CachedArray(T,N)

using InplaceArrays.Arrays
# Create an empty CachedArray
a = CachedArray(Float64,2)
# Resize to new shape (2,3)
setsize!(a,(2,3))
size(a)
# output
(2, 3)

source

InplaceArrays.Arrays.CachedArray — Method.

CachedArray(a)

source

InplaceArrays.Arrays.CachedArray — Method.

CachedArray(T, N)

source

InplaceArrays.Arrays.setsize! — Function.

setsize!(a, s)

Changes the size of the CachedArray a to the size described the the tuple s.

Warning

After calling setsize!, the array can store uninitialized values.

source

InplaceArrays.Arrays.CachedMatrix — Type.

const CachedMatrix{T,A} = CachedArray{T,2,A}

source

InplaceArrays.Arrays.CachedVector — Type.

const CachedVector{T,A} = CachedArray{T,1,A}

source