Extended AbstractArray interface

Extended AbstractArray interface

When implementing new array types, we encounter a similar problem than when implementing some functions: It can be needed some scratch data (e.g., allocating the output), when recovering an item from an array (typically if the array elements are mutable or non-isbits objects, e.g., for "lazy" array of arrays). Here, we adopt the same solution as for functors: the user provides the scratch data. However, the Julia array interface does not support this approach. When calling a[i], in order to get the element with index i in array a, there is no extra argument for the scratch data. In order to circumvent this problem, we add new methods to the AbstractArray interface of Julia. We provide default implementations to the new methods, so that any AbstractArray can be used with the extended interface. The most important among the new methods is getindex!, which allows to recover an item in the array by passing some scratch data. The new mehtods are listed below.

New functions

InplaceArrays.Arrays.getindex! — Function.

getindex!(cache,a::AbstractArray,i...)

Returns the item of the array a associated with index i by (possibly) using the scratch data passed in the cache object.

It defaults to

getindex!(cache,a::AbstractArray,i...) = a[i...]

The cache object is constructed with the array_cache function.

Examples

julia> a = collect(1:4)
4-element Array{Int64,1}:
 1
 2
 3
 4

julia> cache = array_cache(a)

julia> getindex!(cache,a,2)
2

julia> getindex!(cache,a,4)
4

In this example, using the extended interface provides little benefit, but for new array types that need scratch data, efficient implementations of getindex! can make a performance difference by avoiding low granularity allocations.

source

InplaceArrays.Arrays.array_cache — Function.

array_cache(a::AbstractArray)

Returns a cache object to be used in the getindex! function. It defaults to

array_cache(a::T) where T = nothing

for types T such that uses_hash(T) == Val(false), and

function array_cache(a::T) where T
  hash = Dict{UInt,Any}()
  array_cache(hash,a)
end

for types T such that uses_hash(T) == Val(true), see the uses_hash function. In the later case, the type T should implement the following signature:

array_cache(hash::Dict,a::AbstractArray)

where we pass a dictionary (i.e., a hash table) in the first argument. This hash table can be used to test if the object a has already build a cache and re-use it as follows

id = objectid(a)
if haskey(hash,id)
  cache = hash[id] # Reuse cache
else
  cache = ... # Build a new cache depending on your needs
  hash[id] = cache # Register the cache in the hash table
end

This mechanism is needed, e.g., to re-use intermediate results in complex lazy operation trees. In multi-threading computations, a different hash table per thread has to be used in order to avoid race conditions.

source

InplaceArrays.Arrays.uses_hash — Function.

uses_hash(::Type{T}) where T <:AbstractArray

This function is used to specify if the type T uses the hash-based mechanism to reuse caches. It should return either Val(true) or Val(false). It defaults to

uses_hash(::Type{<:AbstractArray}) = Val(false)

Once this function is defined for the type T it can also be called on instances of T.

Examples

julia> uses_hash(Matrix{Float64})
Val{false}()

julia> a = ones(2,3)
2×3 Array{Float64,2}:
 1.0  1.0  1.0
 1.0  1.0  1.0

julia> uses_hash(a)
Val{false}()

source

InplaceArrays.Arrays.testitem — Function.

testitem(a::AbstractArray)

Returns an arbitrary instance of eltype(a). The default returned value is the first entry in the array if length(a)>0 and testvalue(eltype(a)) if length(a)==0 See the testvalue function.

This function is useful to determine the type resulting from applying a given function to the items in the array without calling the Base._return_type function.

Examples

julia> a = collect(1:0)
0-element Array{Int64,1}

julia> ai = testitem(a) # Safely works with empty arrays
0

julia> typeof(sqrt(ai))
Float64

source

InplaceArrays.Arrays.testvalue — Function.

testvalue(::Type{T}) where T

Returns an arbitrary instance of type T. It defaults to zero(T) for non-array types and to an empty array for array types. This function is useful to determine the type returned by a function without calling Base._return_type.

Examples

julia> a = testvalue(Int)
0

julia> b = testvalue(Float64)
0.0

julia> typeof(a + b)
Float64

source

Working with several arrays

InplaceArrays.Arrays.testitems — Function.

testitems(b::AbstractArray...) -> Tuple

Returns a tuple with the result of testitem applied to each of the arrays in b.

source

Creating lazy operation trees

InplaceArrays.Arrays.evaluate_functor_with_arrays — Function.

evaluate_functor_with_arrays(f,a::AbstractArray...)

Returns a (lazy) array representing the evaluation of the given functor f to the entries of the input arrays a. The returned array r is such that r[i] == evaluate_functor(f,a[1][i],a[2][i],...) Items in the resulting array r can be memory-efficiently recovered by using the getindex! function. Note that this function returns a lazy object. The operations are not performed until an entry of the array is retrieved. By applying this function to the result r again (possibly with another functor) we create a lazy operation tree. The underlying implementation is able to reuse intermediate results in this operation tree.

Examples

julia> a = collect(1:5)
5-element Array{Int64,1}:
 1
 2
 3
 4
 5

julia> b = collect(6:10)
5-element Array{Int64,1}:
  6
  7
  8
  9
 10

julia> c = evaluate_functor_with_arrays(+,a,b)
5-element InplaceArrays.Arrays.EvaluatedArray{Int64,1,IndexLinear(),Tuple{Array{Int64,1},Array{Int64,1}},FillArrays.Fill{typeof(+),1,Tuple{Base.OneTo{Int64}}}}:
  7
  9
 11
 13
 15

julia> d = evaluate_functor_with_arrays(*,c,c)
5-element InplaceArrays.Arrays.EvaluatedArray{Int64,1,IndexLinear(),Tuple{InplaceArrays.Arrays.EvaluatedArray{Int64,1,IndexLinear(),Tuple{Array{Int64,1},Array{Int64,1}},FillArrays.Fill{typeof(+),1,Tuple{Base.OneTo{Int64}}}},InplaceArrays.Arrays.EvaluatedArray{Int64,1,IndexLinear(),Tuple{Array{Int64,1},Array{Int64,1}},FillArrays.Fill{typeof(+),1,Tuple{Base.OneTo{Int64}}}}},FillArrays.Fill{typeof(*),1,Tuple{Base.OneTo{Int64}}}}:
  49
  81
 121
 169
 225

source

Working with arrays of functors

InplaceArrays.Arrays.evaluate_array_of_functors — Function.

evaluate_array_of_functors(f::AbstractArray,a::AbstractArray...)

Examples

julia> g = [+,-,*,mod,max]
5-element Array{Function,1}:
 +  
 -  
 *  
 mod
 max

julia> x = collect(1:5)
5-element Array{Int64,1}:
 1
 2
 3
 4
 5

julia> y = collect(6:10)
5-element Array{Int64,1}:
  6
  7
  8
  9
 10

julia> evaluate_array_of_functors(g,x,y)
5-element InplaceArrays.Arrays.EvaluatedArray{Int64,1,IndexLinear(),Tuple{Array{Int64,1},Array{Int64,1}},Array{Function,1}}:
  7
 -5
 24
  4
 10

source

InplaceArrays.Arrays.compose_functor_with_arrays — Function.

compose_functor_with_arrays(g,f::AbstractArray...)

source

InplaceArrays.Arrays.compose_arrays_of_functors — Function.

compose_arrays_of_functors(g::AbstractArray,f::AbstractArray...)

source