The CellValue interface

The CellValue interface

InplaceArrays.CellValues.CellValue — Type.

abstract type CellValue{T}

Abstract type representing a collection of values of type T associated with the objects (e.g, cells, but also facets, edges, etc.) of a FE mesh.

A CellValue has to be understood as an array plus (possibly) some extra metatada, e.g., the underlying mesh, fe space, information about the current field in multi-field FE computations etc. The simplest version of CellValue, namely PlainCellValue, has no metadata.

All concrete types extending CellValue are assumed to have a field named array that contains an instance of AbstractArray{T} representing the collection of values in the CellValue object.

Concrete implementations of CellValue do not need to be type stable. In particular, the type of the array field does not need to be included as a type parameter in the corresponding CellValue concrete type. This allows to pass CellValue objects around without polluting the stack trace if an error occurs. However, the object stored in the array field has to be type stable. This allows to access items in CellValue objects efficiently via the array field (within a function barrier).

source

Constructors

InplaceArrays.CellValues.CellValue — Method.

CellValue(cv::CellValue,array::AbstractArray)

Creates a CellValue object from the array array and (possibly) the metadata available in cv. By default, the metadata in cv is discarded and a PlainCellValue is returned. However, concrete implementations of the CellValue abstract type can overload this constructor in order to generate a new CellValue object from the given array and the metadata in cv.

source

InplaceArrays.CellValues.CellValue — Method.

CellValue(array::AbstractArray) -> PlainCellValue

Creates a PlainCellValue from the given array

source

InplaceArrays.CellValues.CellValue — Method.

CellValue(value,len::Integer) -> PlainCellValue

Creates a "constant" CellValue object with value value and length len.

source

InplaceArrays.CellValues.CellValue — Method.

CellValue(value,shape::Integer...) -> PlainCellValue

Creates a "constant" CellValue object with value value and length prod(shape). The size of the generated array is equal to shape.

source

Methods delegated to the underlying array

Base.length — Method.

Base.length(cv::CellValue)

Returns the length of the underlying array

source

Default concrete implementations

InplaceArrays.CellValues.PlainCellValue — Type.

struct PlainCellValue{T} <: CellValue{T}
  array::AbstractArray{T}
end

Concrete implementation of CellValue with no meta-data

source

Working with several CellValue objects

InplaceArrays.CellValues.getarrays — Function.

getarrays(cvs::CellValue...) -> Tuple

Returns a tuple with the underlying arrays in the Cellvalue objects cvs.

source

Creating lazy operation trees

InplaceArrays.CellValues.apply — Function.

apply(f,cvs::CellValue...)

Returns a new CellValue object obtained by applying the functor f to the entries of the given CellValue objects cvs.

source

CellValue objects holding numeric data

InplaceArrays.CellValues.CellNumber — Type.

const CellNumber = CellValue{T} where T<:Number

Any CellValue{T} type holding numbers of type T.

source

InplaceArrays.CellValues.CellArray — Type.

const CellArray = CellValue{T} where T<:AbstractArray{S,N} where {S,N}

Any CellValue{T} type holding arrays of type T in each entry.

source

InplaceArrays.CellValues.CellData — Type.

const CellData = CellValue{T} where T<:Union{Number,AbstractArray}

Any CellValue{T} type holding numbers or arrays of type T.

source

The arithmetic operations +, -, * are overloaded for CellNumber, CellArray, and CellData objects.