Gridap.Fields

Gridap.Fields

InplaceArrays.Fields — Module.

This module provides:

An interface for physical fields and basis of physical fields.

Helpers functions to work with fields and basis.

Helpers functions to create lazy operation trees from fields and arrays of fields

The exported names are:

Basis

Field

Point

apply_kernel_to_field

apply_to_field

compose

evaluate

evaluate!

field_cache

field_return_type

gradient

lincomb

pointdim

test_array_of_fields

test_field

valuetype

∇

source

Interface

InplaceArrays.Fields.Field — Type.

abstract type Field{V,D} <: Kernel

Abstract type representing physical field, basis of fields, and other related objects.

D is the number of components of the points where the field can be evaluated.

V has to be a type <:Number if the field returns a number when evaluated at a single point.

V has to be a type <:AbstractArray if the field returns an array when evaluated at a single point. E.g., for basis of fields V is a type <:AbstractVector.

Note that eltype(V) is allowed to be any number type since the actual returned type can depend in general on the particular type of the evaluation point components.

The following functions need to be overloaded for derived types:

evaluate!(cache,f::Field,x::Point)

field_cache(f::Field,x::Point)

The following functions can be also provided optionally

gradient(f::Field)

field_return_type(f::Field,x::Point)

The following vectorized versions can be (optionally) rewritten for specific types to improve performance:

evaluate!(cache,f::Field,x::AbstractVector{<:Point})

field_cache(f::Field,x::AbstractVector{<:Point})

field_return_type(f::Field,x::AbstractVector{<:Point})

The interface can be tested with

test_field

source

InplaceArrays.Fields.Basis — Type.

const Basis = Field{V,D} where {V<:AbstractVector, D}

Alias for the particular case, where the field returns a vector of values.

source

InplaceArrays.Fields.Point — Type.

const Point{D,T} = VectorValue{D,T}

Type representing a point of D dimensions with coordinates of type T

source

InplaceArrays.Fields.evaluate! — Method.

evaluate!(cache,f::Field,x::Point)

Returns the value of field at point x. The value of a field is typically a number or an array. When the value is a vector, the field is in fact a basis.

source

InplaceArrays.Fields.field_cache — Method.

field_cache(f::Field,x::Point)

Returns the cache object needed to evaluate field f at point x.

source

InplaceArrays.Fields.gradient — Method.

gradient(f::Field) -> Field

Returns another field that represents the gradient of the given one

source

InplaceArrays.Fields.∇ — Method.

gradient(f::Field) -> Field

Returns another field that represents the gradient of the given one

source

gradient(k::Kernel,f::Field...)

Returns a field representing the gradient of the field obtained with

apply_kernel_to_field(k,f...)

source

InplaceArrays.Fields.field_return_type — Method.

field_return_type(f::Field,x::Point)

Computes the type obtained when evaluating field f at point x.

source

InplaceArrays.Fields.evaluate! — Method.

evaluate!(cache,f::Field,x::AbstractVector{<:Point})

Vectorized version of evaluate!(f::Field,x::Point).

For fields f with valuetype(f)<:Number, it returns a vector with the value of f at each of the point in x.

For Fields f with valuetype(f)<:AbstractArray, it returns an array with one dimension more than the value of the field. E.g., for basis, it should return a matrix. The axis associated with the points x is the last axis in the resulting array. E.g., for a basis with ndof degrees of freedom, the returned matrix has size (ndof,length(x)) .

source

InplaceArrays.Fields.field_cache — Method.

field_cache(f::Field,x::AbstractVector{<:Point})

Returns the cache object needed to evaluate the field f at the vector of points x by means of the vectorized version of evaluate!.

source

InplaceArrays.Fields.field_return_type — Method.

field_return_type(f::Field,x::AbstractVector{<:Point})

Returns the type of the object obtained when the field f is evaluated at the vector of points x by means of the vectorized version of evaluate!.

source

InplaceArrays.Fields.test_field — Function.

test_field(
  f::Field,
  x::AbstractVector{<:Point},
  v::AbstractArray,cmp=(==);
  grad=nothing)

Function used to test the field interface.

source

Other functions using fields

InplaceArrays.Fields.evaluate — Method.

evaluate(f::Field,x)

Equivalent to

cache = field_cache(f,x)
evaluate!(cache,f,x)

source

InplaceArrays.Fields.valuetype — Method.

valuetype(::Type{Field{T,D}}) where {T,D}

Returns T

source

InplaceArrays.Fields.pointdim — Method.

pointdim(::Type{Field{T,D}}) where {T,D}

Returns D

source

InplaceArrays.Fields.compose — Method.

compose(g::Function,f...)

Returns a new field obtained by composition of function g and the fields f. The value of the resulting field at point x is numerically equivalent to

fx = [ evaluate(fi,x) for fi in f]
apply_kernel(elem(g), fx...)

The gradient of the resulting field evaluated at point x is equivalent to

fx = [ evaluate(fi,x) for fi in f]
apply_kernel(elem(gradient(g)), fx...)

Note that it is needed to overload gradient(::typeof(g)) for the given function g in order to be able to compute the gradient.

As in function apply_kernel_to_field if any of the inputs in f is a number or an array instead of a field it will be treated as a "constant field".

source

InplaceArrays.Fields.lincomb — Method.

lincomb(a::Basis,b::AbstractVector)

Returns a field f with valuetype(f) <: Number obtained by the "linear combination" of the value of the basis a and the vector b. That is, the value of the resulting field f at a point x is defined as

k = contract(outer)
ax = evaluate(a,x)
apply_kernel(k,ax,b)

On the other hand, the gradient of the resulting field is defined as

k = contract(outer)
∇ax = evaluate(gradient(a),x)
apply_kernel(k,∇ax,b)

source

Applying kernels to fields

InplaceArrays.Fields.apply_kernel_to_field — Method.

apply_kernel_to_field(k::Kernel,f...) -> Field

Returns a field obtained by applying the kernel k to the values of the fields in f. That is, the returned field evaluated at a point x provides the value obtained by applying kernel k to the values of the fields f at point x. Formally, the resulting field at a point x is defined as

fx = [ evaluate(fi,x) for fi in f]
apply_kernel(k,fx...)

if any of the inputs in f is a number or an array of numbers instead of a field it will be treated as a "constant field". That is a filed that evaluated at any point x returns always the underlying number or array.

In order to be able to call the gradient function of the resulting field, one needs to define the gradient operator associated with the underlying kernel. This is done by adding a new method gradient(k::Kernel,f::Field...) for each kernel type.

source

InplaceArrays.Fields.gradient — Method.

gradient(k::Kernel,f::Field...)

Returns a field representing the gradient of the field obtained with

apply_kernel_to_field(k,f...)

source

Working with arrays of fields

InplaceArrays.Fields.evaluate — Method.

evaluate(a::AbstractArray{<:Field},x::AbstractArray)

Evaluates the fields in the array a at the locations provided in the array x (which can be an array of points or an array of vectors of points).

The result is numerically equivalent to

map(evaluate,a,x)

source

InplaceArrays.Fields.gradient — Method.

gradient(a::AbstractArray{<:Field})

Returns an array containing the gradients of the fields in the array a. Numerically equivalent to

map(gradient,a)

source

InplaceArrays.Fields.apply_to_field — Method.

apply_to_field(k::Kernel,f::AbstractArray...)

Returns an array of fields numerically equivalent to

map( (x...) -> apply_kernel_to_field(k,x...), f )

source

InplaceArrays.Fields.field_cache — Method.

field_cache(a::AbstractArray{<:Field},x::AbstractArray) -> Tuple

Returns the caches needed to perform the following iteration

ca, cfi, cx = field_cache(a,x)

for i in length(a)
  fi = getindex!(ca,a,i)
  xi = getindex!(cx,x,i)
  fxi = evaluate!(cfi,fi,xi)
end

source

InplaceArrays.Fields.compose — Method.

compose(g::Function,f::AbstractArray...)

Returns an array of fields numerically equivalent to

map( (x...)->compose(g,x...), f...)

source

InplaceArrays.Fields.lincomb — Method.

lincomb(a::AbstractArray{<:Field},b::AbstractArray)

Returns an array of field numerically equivalent to

map(lincomb,a,b)

source

InplaceArrays.Fields.test_array_of_fields — Function.

function test_array_of_fields(
  a::AbstractArray{<:Field},
  x::AbstractArray,
  v::AbstractArray,
  cmp::Function=(==);
  grad = nothing)

Function to test an array of fields.

source