Physical fields

Physical fields

Interface

InplaceArrays.Fields.FieldLike — Type.

abstract type FieldLike{D,T,N}

Abstract type representing either a field ( for N==1) or a basis of fields (for N==2) of value T, evaluable at points with D components.

The following functions need to be overloaded for derived types:

evaluate!

new_cache

return_type(::FieldLike)

num_dofs (Only for CellBasis, i.e. N==2.)

The following functions can optionally be also provided

gradient(f::FieldLike)

The interface can be tested with these functions

test_fieldlike

test_field

test_field_with_gradient

source

InplaceArrays.Fields.Field — Type.

const Field = FieldLike{D,T,1} where {D,T}

source

InplaceArrays.Fields.Basis — Type.

const Basis = FieldLike{D,T,2} where {D,T}

source

InplaceArrays.Fields.Point — Type.

const Point{D,T} = VectorValue{D,T}

Type representing a point of D dimensions with coordinates of type T

source

InplaceArrays.Fields.evaluate! — Function.

evaluate!(cache,f::FieldLike,x::AbstractVector{<:Point}) -> AbstractArray

For Fields it returns an instance of AbstractVector and for Basis and an instance of AbstractMatrix.

source

InplaceArrays.Fields.new_cache — Function.

new_cache(f::FieldLike)

source

InplaceArrays.Inference.return_type — Method.

return_type(::FieldLike) -> DataType

source

InplaceArrays.Fields.gradient — Method.

gradient(f::FieldLike) -> FieldLike

source

InplaceArrays.Fields.num_dofs — Function.

num_dofs(b::Basis) -> Int

source

InplaceArrays.Fields.test_fieldlike — Function.

source

InplaceArrays.Fields.test_field — Function.

source

InplaceArrays.Fields.test_field_with_gradient — Function.

source

InplaceArrays.Fields.test_basis — Function.

source

API

InplaceArrays.Fields.evaluate — Function.

evaluate(cf::CellFieldLike,x::CellPoints) -> CellArray

source

evaluate(f::FieldLike,x::AbstractVector{<:Point})

source

InplaceArrays.Fields.valuetype — Function.

valuetype(::Type) -> DataType

source

InplaceArrays.Fields.pointdim — Function.

pointdim(::Type) -> Int

source

InplaceArrays.Fields.gradtype — Function.

gradtype(::Type) -> DataType

source