Gridap.Fields

Gridap.Fields

InplaceArrays.Fields — Module.

This module defines the interface for physical fields and basis of physical fields. It also provides some helpers to work with fields and basis.

The exported names are:

Basis

Field

FieldLike

Point

evaluate

evaluate!

field_cache

gradient

gradtype

num_dofs

pointdim

test_basis

test_field

test_field_like

valuetype

∇

source

Interface

InplaceArrays.Fields.FieldLike — Type.

abstract type FieldLike

Abstract type representing either a field ( for N==1) or a basis of fields (for N==2) of value T, evaluable at points with D components.

The following functions need to be overloaded for derived types:

evaluate!(cache,f::FieldLike,x)

field_cache(f::FieldLike)

num_dofs(f::Basis)

The following functions can optionally be also provided

gradient(f::FieldLike)

The interface can be tested with these functions

test_field

test_basis

source

InplaceArrays.Fields.Field — Type.

const Field = FieldLike{D,T,1} where {D,T}

source

InplaceArrays.Fields.Basis — Type.

const Basis = FieldLike{D,T,2} where {D,T}

source

InplaceArrays.Fields.Point — Type.

const Point{D,T} = VectorValue{D,T}

Type representing a point of D dimensions with coordinates of type T

source

InplaceArrays.Fields.evaluate! — Method.

evaluate!(cache,f::FieldLike,x) -> AbstractArray

For Fields it returns an instance of AbstractVector and for Basis and an instance of AbstractMatrix.

source

InplaceArrays.Fields.field_cache — Method.

field_cache(f::FieldLike)

source

InplaceArrays.Fields.gradient — Method.

gradient(f::FieldLike)

source

InplaceArrays.Fields.∇ — Method.

gradient(f::FieldLike)

source

InplaceArrays.Fields.num_dofs — Method.

num_dofs(b::Basis) -> Int

source

InplaceArrays.Fields.test_field_like — Function.

test_field_like(
  f::FieldLike{D,T,N},
  x::AbstractVector{<:Point},
  v::AbstractArray,cmp=(==);
  grad=nothing) where {D,T,N}

source

InplaceArrays.Fields.test_field — Function.

source

InplaceArrays.Fields.test_basis — Function.

source

Other functions using fields

InplaceArrays.Fields.evaluate — Function.

evaluate(f::FieldLike,x::AbstractVector{<:Point})

source

InplaceArrays.Fields.valuetype — Function.

valuetype(::Type) -> DataType

source

InplaceArrays.Fields.pointdim — Function.

pointdim(::Type) -> Int

source

InplaceArrays.Fields.gradtype — Function.

gradtype(::Type) -> DataType

source