Gridap.Fields

Gridap.Fields

InplaceArrays.Fields — Module.

This module defines the interface for physical fields and basis of physical fields. It also provides some helpers to work with fields and basis.

The exported names are:

Field

Point

apply_kernel_to_field

apply_to_field

compose

evaluate

evaluate!

field_cache

field_return_type

gradient

pointdim

test_array_of_fields

test_field

valuetype

∇

source

Interface

InplaceArrays.Fields.Field — Type.

abstract type Field{V,D}

Abstract type representing field.

D is the number of components of the points where the field can be evaluated.

V has to be a type <:Number if the field returns a number when evaluated at a single point.

V has to be a type <:AbstractArray if the field returns an array when evaluated at a single point. E.g., for basis of fields V is a type <:AbstractVector.

Note that eltype(V) is allowed to be any number type since the actual returned type can depend in general on the particular type of the evaluation point components.

The following functions need to be overloaded for derived types:

evaluate!(cache,f::Field,x::Point)

field_cache(f::Field,x::Point)

The following functions can optionally be also provided

gradient(f::Field)

field_return_type(f::Field,x::Point)

The following vectorized versions can be (optionally) rewritten for specific types to improve performance:

evaluate!(cache,f::Field,x::AbstractVector{<:Point})

field_cache(f::Field,x::AbstractVector{<:Point})

field_return_type(f::Field,x::AbstractVector{<:Point})

The interface can be tested with

test_field

source

InplaceArrays.Fields.Point — Type.

const Point{D,T} = VectorValue{D,T}

Type representing a point of D dimensions with coordinates of type T

source

InplaceArrays.Fields.evaluate! — Method.

evaluate!(cache,f::Field,x::Point)

Returns the value of field at point x. The value of a field is typically a number or an array. When the value is a vector, the field is informally referred to as a basis.

source

InplaceArrays.Fields.field_cache — Method.

field_cache(f::Field,x::Point)

source

InplaceArrays.Fields.gradient — Method.

gradient(f::Field) -> Field

source

InplaceArrays.Fields.∇ — Method.

gradient(f::Field) -> Field

source

gradient(k,f...)

source

InplaceArrays.Fields.field_return_type — Method.

field_return_type(f::Field,x::Point)

source

InplaceArrays.Fields.evaluate! — Method.

evaluate!(cache,f::Field,x::AbstractVector{<:Point})

source

InplaceArrays.Fields.field_cache — Method.

field_cache(f::Field,x::AbstractVector{<:Point})

source

InplaceArrays.Fields.field_return_type — Method.

field_return_type(f::Field,x::AbstractVector{<:Point})

source

InplaceArrays.Fields.test_field — Function.

test_field(
  f::Fiel,
  x::AbstractVector{<:Point},
  v::AbstractArray,cmp=(==);
  grad=nothing)

source

Other functions using fields

InplaceArrays.Fields.evaluate — Method.

evaluate(f::Field,x::Point)

source

InplaceArrays.Fields.valuetype — Method.

valuetype(::Type{Field{T,D}}) where {T,D}

Returns T

source

InplaceArrays.Fields.compose — Method.

compose(g::Function,f...)

source

Applying kernels to fields

Missing docstring.

Missing docstring for apply_kernel_to_field(k,f::FieldNumberOrArray...). Check Documenter's build log for details.

InplaceArrays.Fields.gradient — Method.

gradient(k,f...)

source

Working with arrays of fields

InplaceArrays.Fields.evaluate — Method.

evaluate(a::AbstractArray{<:Field},x::AbstractArray)

The result is numerically equivalent to

map(evaluate,a,x)

source

InplaceArrays.Fields.gradient — Method.

gradient(a::AbstractArray{<:Field},x::AbstractArray)

Numerically equivalent to

map(gradient,a)

source

InplaceArrays.Fields.apply_to_field — Method.

apply_to_field(k,f::AbstractArray...)

Numerically equivalent to

map( (x...) -> apply_to_field(k,x...), f )

source

InplaceArrays.Fields.field_cache — Method.

field_cache(a::AbstractArray{<:Field},x::AbstractArray) -> Tuple

Returns the caches needed to perform the following iteration

ca, cfi, cx = field_cache(a,x)

for i in length(a)
  fi = getindex!(ca,a,i)
  xi = getindex!(cx,x,i)
  fxi = evaluate!(cfi,fi,xi)
end

source

InplaceArrays.Fields.compose — Method.

compose(g::Function,f::AbstractArray...)

source

InplaceArrays.Fields.test_array_of_fields — Function.

function test_array_of_fields(
  a::AbstractArray{<:Field},
  x::AbstractArray,
  v::AbstractArray,
  cmp::Function=(==);
  grad = nothing)

source