Gridap.ReferenceFEs

Gridap.ReferenceFEs

InplaceArrays.ReferenceFEs — Module.

The exported names are

Dof

ExtrusionPolytope

HEX

HEX_AXIS

LagrangianDofBasis

Polytope

QUAD

SEGMENT

TET

TET_AXIS

TRI

VERTEX

dof_cache

dof_return_type

edge_tangents

evaluate_dof

evaluate_dof!

evaluate_dof_array

facet_normals

facet_orientations

num_dims

num_edges

num_faces

num_facets

polytope_dimrange

polytope_facedims

polytope_faces

polytope_offset

polytope_offsets

polytope_vtkid

polytope_vtknodes

test_dof

test_polytope

vertex_coordinates

vertex_permutations

source

Polytopes

Interface

InplaceArrays.ReferenceFEs.Polytope — Type.

abstract type Polytope{D}

Abstract type representing a polytope (i.e., a polyhedron in arbitrary dimensions). D is the environment dimension (typically, 0, 1, 2, or 3). This type parameter is needed since there are functions in the Polytope interface that return containers with Point{D} objects. We adopt the usual nomenclature for polytope-related objects. All objects in a polytope (from vertices to the polytope itself) are called n-faces or simply faces. The notation n-faces is used only when it is needed to refer to the object dimension n. Otherwise we simply use face. In addition, we say

vertex (pl. vertices): for 0-faces

edge: for 1-faces

facet: for (D-1)-faces

The Polytope interface is defined by overloading the following functions

polytope_faces(p)

polytope_dimrange(p)

Polytope{N}(p,faceid) where N

vertex_coordinates(p)

(==)(a::Polytope{D},b::Polytope{D}) where D

And optionally these ones:

edge_tangents(p)

facet_normals(p)

facet_orientations(p)

vertex_permutations(p)

polytope_vtkid(p)

polytope_vtknodes(p)

The interface can be tested with the function

test_polytope

source

InplaceArrays.ReferenceFEs.polytope_faces — Method.

polytope_faces(p) -> Vector{Vector{Int}}

source

InplaceArrays.ReferenceFEs.polytope_dimrange — Method.

polytope_dimrange(p) -> Vector{UnitRange{Int}}

source

InplaceArrays.ReferenceFEs.Polytope — Method.

Polytope{N}(p,faceid) where N

Returns a Polytope{N} object representing the N-face with id faceid. The value faceid refers to the numeration restricted to the dimension N (it starts with 1 for the first N-face).

source

InplaceArrays.ReferenceFEs.vertex_coordinates — Method.

vertex_coordinates(p) -> Vector{Point{D,Float64}}

source

Base.:== — Method.

(==)(a::Polytope{D},b::Polytope{D}) where D

Note that the operator == returns false by default for polytopes of different dimensions.

source

InplaceArrays.ReferenceFEs.edge_tangents — Method.

edge_tangents(p) -> Vector{VectorValue{D,Float64}}

source

InplaceArrays.ReferenceFEs.facet_normals — Method.

facet_normals(p) -> Vector{VectorValue{D,Float64}}

source

InplaceArrays.ReferenceFEs.facet_orientations — Method.

facet_orientations(p) -> Vector{Int}

source

InplaceArrays.ReferenceFEs.vertex_permutations — Method.

vertex_permutations(p) -> Vector{Vector{Int}}

source

InplaceArrays.ReferenceFEs.polytope_vtkid — Method.

polytope_vtkid(p)

source

InplaceArrays.ReferenceFEs.polytope_vtknodes — Method.

polytope_vtknodes(p)

source

InplaceArrays.ReferenceFEs.test_polytope — Method.

test_polytope(p::Polytope{D}; optional::Bool=false) where D

Function that stresses out the functions in the Polytope interface. with optional=false (the default), only the mandatory functions are tested. With optional=true, the optional functions are also tested except polytope_vtkid nor polytope_vtknodes

source

InplaceArrays.ReferenceFEs.num_dims — Method.

num_dims(::Type{<:Polytope{D}}) where D
num_dims(p::Polytope{D}) where D

Returns D.

source

InplaceArrays.ReferenceFEs.num_faces — Method.

num_faces(p::Polytope)

source

InplaceArrays.ReferenceFEs.num_faces — Method.

num_faces(p::Polytope,dim::Integer)

source

InplaceArrays.ReferenceFEs.num_facets — Method.

num_facets(p::Polytope)

source

InplaceArrays.ReferenceFEs.num_edges — Method.

num_edges(p::Polytope)

source

InplaceArrays.ReferenceFEs.num_vertices — Method.

num_vertices(p::Polytope)

source

InplaceArrays.ReferenceFEs.polytope_facedims — Method.

polytope_facedims(p) -> Vector{Int}

source

InplaceArrays.ReferenceFEs.polytope_offsets — Method.

polytope_offsets(p) -> Vector{Int}

source

InplaceArrays.ReferenceFEs.polytope_offset — Method.

polytope_offset(p,d)

source

InplaceArrays.ReferenceFEs.polytope_faces — Method.

polytope_faces(p,dimfrom,dimto) -> Vector{Vector{Int}}

source

Concrete implementations

InplaceArrays.ReferenceFEs.ExtrusionPolytope — Type.

Extrusion public, other fields private

source

InplaceArrays.ReferenceFEs.Polytope — Method.

source

InplaceArrays.ReferenceFEs.ExtrusionPolytope — Method.

source

InplaceArrays.ReferenceFEs.HEX_AXIS — Constant.

source

InplaceArrays.ReferenceFEs.TET_AXIS — Constant.

source

Pre-defined polytope instances

InplaceArrays.ReferenceFEs.VERTEX — Constant.

source

InplaceArrays.ReferenceFEs.SEGMENT — Constant.

source

InplaceArrays.ReferenceFEs.TRI — Constant.

source

InplaceArrays.ReferenceFEs.QUAD — Constant.

source

InplaceArrays.ReferenceFEs.TET — Constant.

source

InplaceArrays.ReferenceFEs.HEX — Constant.

source

Degrees of freedom

Interface

InplaceArrays.ReferenceFEs.Dof — Type.

abstract type Dof <: Kernel

Abstract type representing a degree of freedom (DOF), a basis of DOFs, and related objects. These different cases are distinguished by the return type obtained when evaluating the Dof object on a Field object. When a dof is evaluated on a physical field, a scalar number is returned. If either the Dof object is a basis of DOFs, or the Field object is a basis of fields, or both objects are bases, then the returned object is an array of scalar numbers. The first dimensions in the resulting array are for the Dof object and the last ones for the Field object. E.g, a basis of nd DOFs evaluated at physical field returns a vector of nd entries. A basis of nd DOFs evaluated at a basis of nf fields returns a matrix of size (nd,nf).

The following functions needs to be overloaded

dof_cache

evaluate_dof!

The following functions can be overloaded optionally

dof_return_type

In most of the cases it is not strictly needed that types that implement this interface inherit from Dof. However, we recommend to inherit from Dof, when possible.

source

InplaceArrays.ReferenceFEs.dof_cache — Method.

source

InplaceArrays.ReferenceFEs.evaluate_dof! — Method.

source

InplaceArrays.ReferenceFEs.dof_return_type — Method.

source

InplaceArrays.ReferenceFEs.test_dof — Method.

source

InplaceArrays.ReferenceFEs.evaluate_dof — Method.

source

InplaceArrays.Fields.evaluate — Method.

source

InplaceArrays.ReferenceFEs.evaluate_dof_array — Method.

source

InplaceArrays.Fields.evaluate — Method.

source

Concrete implementations

InplaceArrays.ReferenceFEs.LagrangianDofBasis — Type.

source

InplaceArrays.ReferenceFEs.LagrangianDofBasis — Method.

source