Gridap.ReferenceFEs

Gridap.ReferenceFEs

InplaceArrays.ReferenceFEs — Module.

The exported names are

Dof

ExtrusionPolytope

GenericRefFE

HEX

HEX_AXIS

INVALID_PERM

LagrangianDofBasis

LagrangianRefFE

PYRAMID

Polytope

QUAD

ReferenceFE

SEGMENT

SerendipityRefFE

TET

TET_AXIS

TRI

VERTEX

WEDGE

compute_face_orders

compute_lagrangian_reffaces

compute_monomial_basis

compute_nodes

compute_own_nodes

compute_own_nodes_permutations

compute_shapefuns

dof_cache

dof_return_type

evaluate_dof

evaluate_dof!

evaluate_dof_array

get_dimrange

get_dof_to_comp

get_dof_to_node

get_dofs

get_edge_tangents

get_face_own_dofids

get_facedims

get_faces

get_facet_normals

get_facet_orientations

get_node_and_comp_to_dof

get_node_coordinates

get_offset

get_offsets

get_own_dofs_permutations

get_polytope

get_prebasis

get_shapefuns

get_vertex_coordinates

get_vertex_permutations

get_vtkid

get_vtknodes

num_dims

num_dofs

num_edges

num_faces

num_facets

num_vertices

test_dof

test_polytope

test_reference_fe

source

Polytopes

Interface

InplaceArrays.ReferenceFEs.Polytope — Type.

abstract type Polytope{D}

Abstract type representing a polytope (i.e., a polyhedron in arbitrary dimensions). D is the environment dimension (typically, 0, 1, 2, or 3). This type parameter is needed since there are functions in the Polytope interface that return containers with Point{D} objects. We adopt the usual nomenclature for polytope-related objects. All objects in a polytope (from vertices to the polytope itself) are called n-faces or simply faces. The notation n-faces is used only when it is needed to refer to the object dimension n. Otherwise we simply use face. In addition, we say

vertex (pl. vertices): for 0-faces

edge: for 1-faces

facet: for (D-1)-faces

The Polytope interface is defined by overloading the following functions

get_faces(p::Polytope)

get_dimrange(p::Polytope)

Polytope{N}(p::Polytope,faceid::Integer) where N

get_vertex_coordinates(p::Polytope)

(==)(a::Polytope{D},b::Polytope{D}) where D

And optionally these ones:

get_edge_tangents(p::Polytope)

get_facet_normals(p::Polytope)

get_facet_orientations(p::Polytope)

get_vertex_permutations(p::Polytope)

get_vtkid(p::Polytope)

get_vtknodes(p::Polytope)

The interface can be tested with the function

test_polytope

source

InplaceArrays.ReferenceFEs.get_faces — Method.

get_faces(p::Polytope) -> Vector{Vector{Int}}

Given a polytope p the function returns a vector of vectors defining the incidence relation of the faces in the polytope.

Each face in the polytope receives a unique integer id. The id 1 is assigned to the first 0-face. Consecutive increasing ids are assigned to the other 0-faces, then to 1-faces, and so on. The polytope itself receives the largest id which coincides with num_faces(p). For a face id iface, get_faces(p)[iface] is a vector of face ids, corresponding to the faces that are incident with the face labeled with iface. That is, faces that are either on its boundary or the face itself. In this vector of incident face ids, faces are ordered by dimension, starting with 0-faces. Within each dimension, the labels are ordered in a consistent way with the polyope object for the face iface itself.

Examples

using InplaceArrays.ReferenceFEs

faces = get_faces(SEGMENT)
println(faces)

# output
Array{Int64,1}[[1], [2], [1, 2, 3]]

The constant SEGMENT is bound to a predefined instance of polytope that represents a segment. The face labels associated with a segment are [1,2,3], being 1 and 2 for the vertices and 3 for the segment itself. In this case, this function returns the vector of vectors [[1],[2],[1,2,3]] meaning that vertex 1 is incident with vertex 1 (idem for vertex 2), and that the segment (id 3) is incident with the vertices 1 and 2 and the segment itself.

source

InplaceArrays.ReferenceFEs.get_dimrange — Method.

get_dimrange(p::Polytope) -> Vector{UnitRange{Int}}

Given a polytope p it returns a vector of ranges. The entry d+1 in this vector contains the range of face ids for the faces of dimension d.

Examples

using InplaceArrays.ReferenceFEs

ranges = get_dimrange(SEGMENT)
println(ranges)

# output
UnitRange{Int64}[1:2, 3:3]

Face ids for the vertices in the segment range from 1 to 2 (2 vertices), the face ids for edges in the segment range from 3 to 3 (only one edge with id 3).

source

InplaceArrays.ReferenceFEs.Polytope — Method.

Polytope{N}(p::Polytope,faceid::Integer) where N

Returns a Polytope{N} object representing the "reference" polytope of the N-face with id faceid. The value faceid refers to the numeration restricted to the dimension N (it starts with 1 for the first N-face).

source

InplaceArrays.ReferenceFEs.get_vertex_coordinates — Method.

get_vertex_coordinates(p::Polytope) -> Vector{Point{D,Float64}}

Given a polytope p return a vector of points representing containing the coordinates of the vertices.

source

Base.:== — Method.

(==)(a::Polytope{D},b::Polytope{D}) where D

Returns true if the polytopes a and b are equivalent. Otherwise, it returns false. Note that the operator == returns false by default for polytopes of different dimensions. Thus, this function has to be overloaded only for the case of polytopes a and b of same dimension.

source

InplaceArrays.ReferenceFEs.get_edge_tangents — Method.

get_edge_tangents(p::Polytope) -> Vector{VectorValue{D,Float64}}

Given a polytope p, returns a vector of VectorValue objects representing the unit tangent vectors to the polytope edges.

source

InplaceArrays.ReferenceFEs.get_facet_normals — Method.

get_facet_normals(p::Polytope) -> Vector{VectorValue{D,Float64}}

Given a polytope p, returns a vector of VectorValue objects representing the unit outward normal vectors to the polytope facets.

source

InplaceArrays.ReferenceFEs.get_facet_orientations — Method.

get_facet_orientations(p::Polytope) -> Vector{Int}

Given a polytope p returns a vector of integers of length num_facets(p). Facets, whose vertices are ordered consistently with the outwards normal vector, receive value 1 in this vector. Otherwise, facets receive value -1.

source

InplaceArrays.ReferenceFEs.get_vertex_permutations — Method.

get_vertex_permutations(p::Polytope) -> Vector{Vector{Int}}

Given a polytope p, returns a vector of vectors containing all admissible permutations of the polytope vertices. An admissible permutation is one such that, if the vertices of the polytope are re-labeled according to this permutation, the resulting polytope preserves the shape of the original one.

Examples

using InplaceArrays.ReferenceFEs

perms = get_vertex_permutations(SEGMENT)
println(perms)

# output
Array{Int64,1}[[1, 2], [2, 1]]

The first admissible permutation for a segment is [1,2],i.e., the identity. The second one is [2,1], i.e., the first vertex is relabeled as 2 and the second vertex is relabeled as 1.

source

InplaceArrays.ReferenceFEs.get_vtkid — Method.

get_vtkid(p::Polytope) -> Int

Given a polytope p, returns an integer with its vtk identifier. Overloading of this function is needed only in order to visualize the underlying polytope with Paraview.

source

InplaceArrays.ReferenceFEs.get_vtknodes — Method.

get_vtknodes(p::Polytope) -> Vector{Int}

Given a polytope p, returns a vector of integers representing a permutation of the polytope vertices required to relabel the vertices according the criterion adopted in Paraview. Overloading of this function is needed only in order to visualize the underlying polytope with Paraview.

source

InplaceArrays.ReferenceFEs.test_polytope — Method.

test_polytope(p::Polytope{D}; optional::Bool=false) where D

Function that stresses out the functions in the Polytope interface. It tests whether the function in the polytope interface are defined for the given object, and whether they return objects of the expected type. With optional=false (the default), only the mandatory functions are checked. With optional=true, the optional functions are also tested except get_vtkid and get_vtknodes.

source

InplaceArrays.ReferenceFEs.num_dims — Method.

num_dims(::Type{<:Polytope{D}}) where D
num_dims(p::Polytope{D}) where D

Returns D.

source

InplaceArrays.ReferenceFEs.num_faces — Method.

num_faces(p::Polytope)

Returns the total number of faces in polytope p (from vertices to the polytope itself).

source

InplaceArrays.ReferenceFEs.num_faces — Method.

num_faces(p::Polytope,dim::Integer)

Returns the number of faces of dimension dim in polytope p.

source

InplaceArrays.ReferenceFEs.num_facets — Method.

num_facets(p::Polytope)

Returns the number of facets in the polytope p.

source

InplaceArrays.ReferenceFEs.num_edges — Method.

num_edges(p::Polytope)

Returns the number of edges in the polytope p.

source

InplaceArrays.ReferenceFEs.num_vertices — Method.

num_vertices(p::Polytope)

Returns the number of vertices in the polytope p.

source

InplaceArrays.ReferenceFEs.get_facedims — Method.

get_facedims(p::Polytope) -> Vector{Int}

Given a polytope p, returns a vector indicating the dimension of each face in the polytope

Examples

using InplaceArrays.ReferenceFEs

dims = get_facedims(SEGMENT)
println(dims)

# output
[0, 0, 1]

The first two faces in the segment (the two vertices) have dimension 0 and the third face (the segment itself) has dimension 1

source

InplaceArrays.ReferenceFEs.get_offsets — Method.

get_offsets(p::Polytope) -> Vector{Int}

Given a polytope p, it returns a vector of integers. The position in the d+1 entry in this vector is the offset that transforms a face id in the global numeration in the polytope to the numeration restricted to faces to dimension d.

Examples

using InplaceArrays.ReferenceFEs

offsets = get_offsets(SEGMENT)
println(offsets)

# output
[0, 2]

source

InplaceArrays.ReferenceFEs.get_offset — Method.

get_offset(p::Polytope,d::Integer)

Equivalent to get_offsets(p)[d+1].

source

InplaceArrays.ReferenceFEs.get_faces — Method.

get_faces(p::Polytope,dimfrom::Integer,dimto::Integer) -> Vector{Vector{Int}}

For dimfrom >= dimto returns a vector that for each face of dimension dimfrom stores a vector of the ids of faces of dimension dimto on its boundary.

For dimfrom < dimto returns a vector that for each face of dimfrom stores a vector of the face ids of faces of dimension dimto that touch it.

The numerations used in this funcitons are the ones restricted to each dimension.

using InplaceArrays.ReferenceFEs

edge_to_vertices = get_faces(QUAD,1,0)
println(edge_to_vertices)

vertex_to_edges_around = get_faces(QUAD,0,1)
println(vertex_to_edges_around)

# output
Array{Int64,1}[[1, 2], [3, 4], [1, 3], [2, 4]]
Array{Int64,1}[[1, 3], [1, 4], [2, 3], [2, 4]]

source

Extrusion polytopes

InplaceArrays.ReferenceFEs.ExtrusionPolytope — Type.

struct ExtrusionPolytope{D} <: Polytope{D}
  extrusion::Point{D,Int}
  # + private fields
end

Concrete type for polytopes that can be represented with an "extrusion" tuple. The underlying extrusion is available in the field extrusion. Instances of this type can be obtained with the constructors

Polytope(extrusion::Int...)

ExtrusionPolytope(extrusion::Int...)

source

InplaceArrays.ReferenceFEs.ExtrusionPolytope — Method.

ExtrusionPolytope(extrusion::Int...)

Generates an ExtrusionPolytope from the tuple extrusion. The values in extrusion are either equal to the constant HEX_AXIS or the constant TET_AXIS.

Examples

Creating a quadrilateral, a triangle, and a wedge

using InplaceArrays.ReferenceFEs

quad = ExtrusionPolytope(HEX_AXIS,HEX_AXIS)

tri = ExtrusionPolytope(TET_AXIS,TET_AXIS)

wedge = ExtrusionPolytope(TET_AXIS,TET_AXIS,HEX_AXIS)

println(quad == QUAD)
println(tri == TRI)
println(wedge == WEDGE)

# output
true
true
true

source

InplaceArrays.ReferenceFEs.Polytope — Method.

Polytope(extrusion::Int...)

Equivalent to ExtrusionPolytope(extrusion...)

source

InplaceArrays.ReferenceFEs.HEX_AXIS — Constant.

Constant to be used in order to indicate a hex-like extrusion axis.

source

InplaceArrays.ReferenceFEs.TET_AXIS — Constant.

Constant to be used in order to indicate a tet-like extrusion axis.

source

Pre-defined polytope instances

InplaceArrays.ReferenceFEs.VERTEX — Constant.

const VERTEX = Polytope()

source

InplaceArrays.ReferenceFEs.SEGMENT — Constant.

const SEGMENT = Polytope(HEX_AXIS)

source

InplaceArrays.ReferenceFEs.TRI — Constant.

const TRI = Polytope(TET_AXIS,TET_AXIS)

source

InplaceArrays.ReferenceFEs.QUAD — Constant.

const QUAD = Polytope(HEX_AXIS,HEX_AXIS)

source

InplaceArrays.ReferenceFEs.TET — Constant.

const TET = Polytope(TET_AXIS,TET_AXIS,TET_AXIS)

source

InplaceArrays.ReferenceFEs.HEX — Constant.

const HEX = Polytope(HEX_AXIS,HEX_AXIS,HEX_AXIS)

source

InplaceArrays.ReferenceFEs.WEDGE — Constant.

const WEDGE = Polytope(TET_AXIS,TET_AXIS,HEX_AXIS)

source

InplaceArrays.ReferenceFEs.PYRAMID — Constant.

const PYRAMID = Polytope(HEX_AXIS,HEX_AXIS,TET_AXIS)

source

Degrees of freedom

Interface

InplaceArrays.ReferenceFEs.Dof — Type.

abstract type Dof <: Kernel

Abstract type representing a degree of freedom (DOF), a basis of DOFs, and related objects. These different cases are distinguished by the return type obtained when evaluating the Dof object on a Field object. See function evaluate_dof! for more details.

The following functions needs to be overloaded

dof_cache

evaluate_dof!

The following functions can be overloaded optionally

dof_return_type

The interface is tested with

test_dof

In most of the cases it is not strictly needed that types that implement this interface inherit from Dof. However, we recommend to inherit from Dof, when possible.

source

InplaceArrays.ReferenceFEs.evaluate_dof! — Method.

evaluate_dof!(cache,dof,field)

Evaluates the dof dof with the field field. It can return either an scalar value or an array of scalar values depending the case. The cache object is computed with function dof_cache.

When a mathematical dof is evaluated on a physical field, a scalar number is returned. If either the Dof object is a basis of DOFs, or the Field object is a basis of fields, or both objects are bases, then the returned object is an array of scalar numbers. The first dimensions in the resulting array are for the Dof object and the last ones for the Field object. E.g, a basis of nd DOFs evaluated at physical field returns a vector of nd entries. A basis of nd DOFs evaluated at a basis of nf fields returns a matrix of size (nd,nf).

source

InplaceArrays.ReferenceFEs.dof_cache — Method.

dof_cache(dof,field)

Returns the cache needed to call evaluate_dof!(cache,dof,field)

source

InplaceArrays.ReferenceFEs.dof_return_type — Method.

dof_return_type(dof,field)

Returns the type for the value obtained with evaluating dof with field.

It defaults to

typeof(evaluate_dof(dof,field))

source

InplaceArrays.ReferenceFEs.test_dof — Method.

test_dof(dof,field,v,comp::Function=(==))

Test that the Dof interface is properly implemented for object dof. It also checks if the object dof when evaluated at the field field returns the same value as v. Comparison is made with the comp function.

source

InplaceArrays.ReferenceFEs.evaluate_dof — Method.

evaluate_dof(dof,field)

Equivalent to

cache = dof_cache(dof,field)
evaluate_dof!(cache,dof,field)

source

InplaceArrays.Fields.evaluate — Method.

evaluate(dof::Dof,field)

Equivalent to evaluate_dof(dof,field).

source

Working with arrays of DOFs

InplaceArrays.ReferenceFEs.evaluate_dof_array — Method.

evaluate_dof_array(dof::AbstractArray,field::AbstractArray)

Evaluates the Dof objects in the array dof at the Field objects at the array field element by element.

The result is numerically equivalent to

map(evaluate_dof, dof, field)

but it is described with a more memory-friendly lazy type.

source

InplaceArrays.Fields.evaluate — Method.

evaluate(dof::AbstractArray{<:Dof},field::AbstractArray)

Equivalent to evaluate_dof_array(dof,field)

source

Lagrangian dof bases

InplaceArrays.ReferenceFEs.LagrangianDofBasis — Type.

struct LagrangianDofBasis{P,V} <: Dof
  nodes::Vector{P}
  dof_to_node::Vector{Int}
  dof_to_comp::Vector{Int}
  node_and_comp_to_dof::Vector{V}
end

Type that implements a Lagrangian dof basis.

Fields:

nodes::Vector{P} vector of points (P<:Point) storing the nodal coordinates

node_and_comp_to_dof::Vector{V} vector such that node_and_comp_to_dof[node][comp] returns the dof associated with node node and the component comp in the type V.

dof_to_node::Vector{Int} vector of integers such that dof_to_node[dof] returns the node id associated with dof id dof.

dof_to_comp::Vector{Int} vector of integers such that dof_to_comp[dof] returns the component id associated with dof id dof.

source

InplaceArrays.ReferenceFEs.LagrangianDofBasis — Method.

LagrangianDofBasis(::Type{T},nodes::Vector{<:Point}) where T

Creates a LagrangianDofBasis for fields of value type T associated with the vector of nodal coordinates nodes.

source

Reference Finite Elements

Interface

InplaceArrays.ReferenceFEs.ReferenceFE — Type.

abstract type ReferenceFE{D}

Abstract type representing a Reference finite element. D is the underlying coordinate space dimension. We follow the Ciarlet definition. A reference finite element is defined by a polytope (cell topology), a basis of an interpolation space of top of this polytope (denoted here as the prebasis), and a basis of the dual of this space (i.e. the degrees of freedom). From this information one can compute the shape functions (i.e, the canonical basis of w.r.t. the degrees of freedom) with a simple change of basis. In addition, we also encode in this type information about how the interpolation space in a reference finite element is "glued" with neighbors in order to build conforming cell-wise spaces.

The ReferenceFE interface is defined by overloading these methods:

num_dofs(reffe::ReferenceFE)

get_polytope(reffe::ReferenceFE)

get_prebasis(reffe::ReferenceFE)

get_dofs(reffe::ReferenceFE)

get_face_own_dofids(reffe::ReferenceFE)

and optionally these ones:

ReferenceFE{N}(reffe::ReferenceFE,nfaceid::Integer) where N

get_own_dofs_permutations(reffe::ReferenceFE)

source

InplaceArrays.ReferenceFEs.num_dofs — Method.

num_dofs(reffe::ReferenceFE) -> Int

Returns the number of DOFs.

source

InplaceArrays.ReferenceFEs.get_polytope — Method.

get_polytope(reffe::ReferenceFE) -> Polytope

Returns the underlying polytope object.

source

InplaceArrays.ReferenceFEs.get_prebasis — Method.

get_prebasis(reffe::ReferenceFE) -> Field

Returns the underlying prebasis encoded as a Field object.

source

InplaceArrays.ReferenceFEs.get_dofs — Method.

get_dofs(reffe::ReferenceFE) -> Dof

Returns the underlying dof basis encoded in a Dof object.

source

InplaceArrays.ReferenceFEs.get_face_own_dofids — Method.

get_face_own_dofids(reffe::ReferenceFE) -> Vector{Vector{Int}}

source

InplaceArrays.ReferenceFEs.ReferenceFE — Method.

ReferenceFE{N}(reffe::ReferenceFE,nfaceid::Integer) where N

Returns a reference FE obtained by the restriction of the given one to the face with nfaceid within dimension N.

source

InplaceArrays.ReferenceFEs.get_own_dofs_permutations — Method.

get_own_dofs_permutations(reffe::ReferenceFE) -> Vector{Vector{Int}}

Returns a vector of vectors indicating how the dofs owned by the reference fe have to be relabeled when the vertices of the underlying polytope are relabeled according the permutations in

polytope = get_polytope(reffe)
vertex_perms = get_vertex_permutations(polytope)

That is, if the vertices are relabeled with the permutation number iperm

oldvertex_to_newvertex = vertex_perms[iperm]

The dofs are relabeled as

olddof_to_newdof = get_own_dofs_permutations(reffe)[iperm]

Note that in some cases, a valid relabeling of the dofs for a corresponding relabeling of the vertices does not exist (i.e., lagrangian FEs with anisotropic order). For these permutations the corresponding vector in get_own_dofs_permutations(reffe) has to be filled with the constant value INVALID_PERM.

Examples

Relabeling of the dofs owned by the 4-th order segment

using InplaceArrays.ReferenceFEs

order = 4
seg5 = LagrangianRefFE(Float64,SEGMENT,order)

vertexperms = get_vertex_permutations(SEGMENT)
dofperms = get_own_dofs_permutations(seg5)

println(vertexperms)
println(dofperms)

# output
Array{Int64,1}[[1, 2], [2, 1]]
Array{Int64,1}[[1, 2, 3], [3, 2, 1]]

Note that when the segment is flipped (the second permutation) the first owned dof becomes the third owned dof, the second owned dof becomes the second one and the third becomes the first one as one would expect.

Relabeling of the dofs owned by an anisotropic quadrilateral

using InplaceArrays.ReferenceFEs

orders = (2,3)
quad12 = LagrangianRefFE(Float64,QUAD,orders)

vertexperms = get_vertex_permutations(QUAD)
dofperms = get_own_dofs_permutations(quad12)

println(vertexperms)
println(dofperms)
println(INVALID_PERM)

# output
Array{Int64,1}[[1, 2, 3, 4], [1, 3, 2, 4], [2, 1, 4, 3], [2, 4, 1, 3], [3, 1, 4, 2], [3, 4, 1, 2], [4, 2, 3, 1], [4, 3, 2, 1]]
Array{Int64,1}[[1, 2], [0, 0], [1, 2], [0, 0], [0, 0], [2, 1], [0, 0], [2, 1]]
0

Note that not all permutations are valid in this case.

source

InplaceArrays.ReferenceFEs.INVALID_PERM — Constant.

Constant of type Int used to signal that a permutation is not valid.

source

InplaceArrays.ReferenceFEs.get_shapefuns — Method.

get_shapefuns(reffe::ReferenceFE) -> Field

Returns the basis of shape functions (i.e. the canonical basis) associated with the reference FE. The result is encoded as a Field object.

source

InplaceArrays.ReferenceFEs.compute_shapefuns — Method.

compute_shapefuns(dofs,prebasis)

Helper function used to compute the shape function basis associated with the dof basis dofs and the basis prebasis.

It is equivalent to

change = inv(evaluate(dofs,prebasis))
change_basis(prebasis,change)

source

InplaceArrays.ReferenceFEs.num_dims — Method.

num_dims(::Type{<:ReferenceFE{D}}) where D
num_dims(reffe::ReferenceFE{D}) where D

Returns D.

source

InplaceArrays.ReferenceFEs.test_reference_fe — Method.

test_reference_fe(reffe::ReferenceFE{D};optional::Bool=false) where D

Test if the methods in the ReferenceFE interface are defined for the object reffe. If optional=false (the default) only the mandatory methods are checked. Otherwise, the optional methods are also tested.

source

Generic reference elements

InplaceArrays.ReferenceFEs.GenericRefFE — Type.

struct GenericRefFE{D} <: ReferenceFE{D}
  polytope::Polytope{D}
  prebasis::Field
  dofs::Dof
  face_own_dofids::Vector{Vector{Int}}
  shapefuns::Field
  ndofs::Int
  own_dofs_permutations::Vector{Vector{Int}}
  reffaces
end

This type is a materialization of the ReferenceFE interface. That is, it is a struct that stores the values of all abstract methods in the ReferenceFE interface. This type is useful to build reference FEs from the underlying ingredients without the need to create a new type.

Note that some fields in this struct are type unstable deliberately in order to simplify the type signature. Don't access them in computationally expensive functions, instead extract the required fields before and pass them to the computationally expensive function.

source

InplaceArrays.ReferenceFEs.GenericRefFE — Method.

GenericRefFE(
  polytope::Polytope{D},
  prebasis::Field,
  dofs::Dof,
  face_own_dofids::Vector{Vector{Int}};
  shapefuns::Field,
  ndofs::Int,
  own_dofs_permutations::Vector{Vector{Int}},
  reffaces) where D

Constructs a GenericRefFE object with the provided data. Positional arguments are mandatory. All Key-word ones are optional.

source

Lagrangian reference elements

InplaceArrays.ReferenceFEs.LagrangianRefFE — Type.

struct LagrangianRefFE{D} <: ReferenceFE{D}
  data::GenericRefFE{D}
  face_own_nodeids::Vector{Vector{Int}}
  own_nodes_permutations::Vector{Vector{Int}},
end

Type representing a Lagrangian finite element. In addition to all the information provided by a ReferenceFE, this type also provides "node-based" information in the following fields:

face_own_nodeids::Vector{Vector{Int}}: nodes owned by each face

own_nodes_permutations::Vector{Vector{Int}}: permutations of the nodes when the vertices are permuted

For this type

get_dofs(reffe) returns a LagrangianDofBasis

get_prebasis(reffe) returns a MonomialBasis

ReferenceFE{N}(reffe,faceid) where N returns a LagrangianRefFE{N}

source

InplaceArrays.ReferenceFEs.LagrangianRefFE — Method.

LagrangianRefFE(
  polytope::Polytope{D},
  prebasis::MonomialBasis,
  dofs::LagrangianDofBasis,
  face_own_nodeids::Vector{Vector{Int}},
  own_nodes_permutations::Vector{Vector{Int}},
  reffaces...) where D

Low level (inner) constructor of LagrangianRefFE.

source

InplaceArrays.ReferenceFEs.LagrangianRefFE — Method.

LagrangianRefFE(::Type{T},p::Polytope,orders) where T
LagrangianRefFE(::Type{T},p::Polytope,order::Int) where T

Builds a LagrangianRefFE object on top of the given polytope. T is the type of the value of the approximation space (e.g., T=Float64 for scalar-valued problems, T=VectorValue{N,Float64} for vector-valued problems with N components). The arguments order or orders are for the polynomial order of the resulting space, which allows isotropic or anisotropic orders respectively (provided that the cell topology allows the given anisotropic order). The argument orders should be an indexable collection of D integers (e.g., a tuple or a vector), being D the number of space dimensions.

In order to be able to use this function, the type of the provided polytope p has to implement the following additional methods. They have been implemented for ExtrusionPolytope in the library. They need to be implemented for new polytope types in order to build Lagangian reference elements on top of them.

compute_monomial_basis(::Type{T},p::Polytope,orders) where T

compute_own_nodes(p::Polytope,orders)

compute_face_orders(p::Polytope,face::Polytope,iface::Int,orders)

The following methods are also used in the construction of the LagrangianRefFE object. A default implementation of them is available in terms of the three previous methods. However, the user can also implement them for new polytope types increasing customization possibilities.

compute_nodes(p::Polytope,orders)

compute_own_nodes_permutations(p::Polytope, interior_nodes)

compute_lagrangian_reffaces(::Type{T},p::Polytope,orders) where T

source

InplaceArrays.ReferenceFEs.compute_monomial_basis — Method.

compute_monomial_basis(::Type{T},p::Polytope,orders) where T -> MonomialBasis

Returns the monomial basis of value type T and order per direction described by orders on top of the polytope p.

source

InplaceArrays.ReferenceFEs.compute_own_nodes — Method.

compute_own_nodes(p::Polytope{D},orders) where D -> Vector{Point{D,Float64}}

Returns the coordinates of the nodes owned by the interior of the polytope associated with a Lagrangian space with the order per direction described by orders.

source

InplaceArrays.ReferenceFEs.compute_face_orders — Method.

compute_face_orders(p::Polytope,face::Polytope,iface::Int,orders)

Returns a vector or a tuple with the order per direction at the face face of the polytope p when restricting the order per direction orders to this face. iface is the face id of face in the numeration restricted to the face dimension.

source

InplaceArrays.ReferenceFEs.compute_nodes — Method.

compute_nodes(p::Polytope,orders)

When called

node_coords, face_own_nodeids = compute_nodes(p,orders)

Returns node_coords, the nodal coordinates of all the Lagrangian nodes associated with the order per direction orders, and face_own_nodeids, being a vector of vectors indicating which nodes are owned by each of the faces of the polytope p.

source

InplaceArrays.ReferenceFEs.compute_own_nodes_permutations — Method.

compute_own_nodes_permutations(
  p::Polytope, own_nodes_coordinates) -> Vector{Vector{Int}}

Returns a vector of vectors with the permutations of the nodes owned by the interior of the polytope.

source

InplaceArrays.ReferenceFEs.compute_lagrangian_reffaces — Method.

compute_lagrangian_reffaces(::Type{T},p::Polytope,orders) where T

Returns a tuple of length D being the number of space dimensions. The entry d+1 of this tuple contains a vector of LagrangianRefFE one for each face of dimension d on the boundary of the polytope.

source

InplaceArrays.ReferenceFEs.get_node_coordinates — Method.

get_node_coordinates(reffe::LagrangianRefFE) -> Vector{Point{D,Float64}}

Returns the nodal coordinates of the underlying LagrangianDofBasis.

source

InplaceArrays.ReferenceFEs.get_dof_to_node — Method.

get_dof_to_node(reffe::LagrangianRefFE) -> Vector{Int}

Returns the field dof_to_node of the underlying LagrangianDofBasis.

source

InplaceArrays.ReferenceFEs.get_dof_to_comp — Method.

get_dof_to_comp(reffe::LagrangianRefFE) -> Vector{Int}

Returns the field dof_to_comp of the underlying LagrangianDofBasis.

source

InplaceArrays.ReferenceFEs.get_node_and_comp_to_dof — Method.

get_node_and_comp_to_dof(reffe::LagrangianRefFE) -> Vector

Returns the field node_and_comp_to_dof of the underlying LagrangianDofBasis.

source

Serendipity reference elements

InplaceArrays.ReferenceFEs.SerendipityRefFE — Function.

SerendipityRefFE(::Type{T},p::Polytope,order::Int) where T

Returns an instance of LagrangianRefFE, whose underlying approximation space is the serendipity space of order order. Implemented for order from 1 to 4. The type of the polytope p has to implement all the queries detailed in the constructor LagrangianRefFE(::Type{T},p::Polytope{D},orders) where {T,D} plus this additional method:

is_serendipity_compatible(p::Polytope)

Examples

using InplaceArrays.ReferenceFEs

order = 2
reffe = SerendipityRefFE(Float64,QUAD,order)

println( num_dofs(reffe) )

# output
8

source

InplaceArrays.ReferenceFEs.is_serendipity_compatible — Method.

is_serendipity_compatible(p::Polytope) -> Bool

Returns true if the polytope p is compatible with the serendipity space (i.e., it is a n-cube). This method is implemented for ExtrusionPolytope, and should be implemented by new polytope types if they are to be used to build serendipity spaces.

source