Gridap.ReferenceFEs

Gridap.ReferenceFEs

InplaceArrays.ReferenceFEs — Module.

The exported names are

Dof

LagrangianDofBasis

dof_cache

dof_return_type

evaluate_dof

evaluate_dof!

evaluate_dof_array

test_dof

source

InplaceArrays.ReferenceFEs.Dof — Type.

abstract type Dof <: Kernel

Abstract type representing a degree of freedom (DOF), a basis of DOFs, and related objects. These different cases are distinguished by the return type obtained when evaluating the Dof object on a Field object. When a dof is evaluated on a physical field, a scalar number is returned. If either the Dof object is a basis of DOFs, or the Field object is a basis of fields, or both objects are bases, then the returned object is an array of scalar numbers. The first dimensions in the resulting array are for the Dof object and the last ones for the Field object. E.g, a basis of nd DOFs evaluated at physical field returns a vector of nd entries. A basis of nd DOFs evaluated at a basis of nf fields returns a matrix of size (nd,nf).

The following functions needs to be overloaded

dof_cache

evaluate_dof!

The following functions can be overloaded optionally

dof_return_type

In most of the cases it is not strictly needed that types that implement this interface inherit from Dof. However, we recommend to inherit from Dof, when possible.

source

InplaceArrays.ReferenceFEs.LagrangianDofBasis — Type.

source

InplaceArrays.ReferenceFEs.LagrangianDofBasis — Method.

source

InplaceArrays.ReferenceFEs.dof_cache — Method.

source

InplaceArrays.ReferenceFEs.dof_return_type — Method.

source

InplaceArrays.ReferenceFEs.evaluate_dof! — Method.

source

InplaceArrays.ReferenceFEs.evaluate_dof — Method.

source

InplaceArrays.ReferenceFEs.evaluate_dof_array — Method.

source

InplaceArrays.ReferenceFEs.test_dof — Function.

source

InplaceArrays.Fields.evaluate — Method.

source

InplaceArrays.Fields.evaluate — Method.

source