Gridap.ReferenceFEs

Gridap.ReferenceFEs

InplaceArrays.ReferenceFEs — Module.

The exported names are

Dof

ExtrusionPolytope

HEX

HEX_AXIS

LagrangianDofBasis

Polytope

QUAD

SEGMENT

TET

TET_AXIS

TRI

VERTEX

dof_cache

dof_return_type

edge_tangents

evaluate_dof

evaluate_dof!

evaluate_dof_array

facet_normals

facet_orientations

num_dims

num_edges

num_faces

num_facets

polytope_dimrange

polytope_facedims

polytope_faces

polytope_offset

polytope_offsets

polytope_vtkid

polytope_vtknodes

test_dof

test_polytope

vertex_coordinates

vertex_permutations

source

InplaceArrays.ReferenceFEs.HEX — Constant.

source

InplaceArrays.ReferenceFEs.HEX_AXIS — Constant.

source

InplaceArrays.ReferenceFEs.QUAD — Constant.

source

InplaceArrays.ReferenceFEs.SEGMENT — Constant.

source

InplaceArrays.ReferenceFEs.TET — Constant.

source

InplaceArrays.ReferenceFEs.TET_AXIS — Constant.

source

InplaceArrays.ReferenceFEs.TRI — Constant.

source

InplaceArrays.ReferenceFEs.VERTEX — Constant.

source

InplaceArrays.ReferenceFEs.Dof — Type.

abstract type Dof <: Kernel

Abstract type representing a degree of freedom (DOF), a basis of DOFs, and related objects. These different cases are distinguished by the return type obtained when evaluating the Dof object on a Field object. When a dof is evaluated on a physical field, a scalar number is returned. If either the Dof object is a basis of DOFs, or the Field object is a basis of fields, or both objects are bases, then the returned object is an array of scalar numbers. The first dimensions in the resulting array are for the Dof object and the last ones for the Field object. E.g, a basis of nd DOFs evaluated at physical field returns a vector of nd entries. A basis of nd DOFs evaluated at a basis of nf fields returns a matrix of size (nd,nf).

The following functions needs to be overloaded

dof_cache

evaluate_dof!

The following functions can be overloaded optionally

dof_return_type

In most of the cases it is not strictly needed that types that implement this interface inherit from Dof. However, we recommend to inherit from Dof, when possible.

source

InplaceArrays.ReferenceFEs.ExtrusionPolytope — Type.

Extrusion public, other fields private

source

InplaceArrays.ReferenceFEs.ExtrusionPolytope — Method.

source

InplaceArrays.ReferenceFEs.LagrangianDofBasis — Type.

source

InplaceArrays.ReferenceFEs.LagrangianDofBasis — Method.

source

InplaceArrays.ReferenceFEs.Polytope — Type.

D is the environment dimension (typically, 0, 1, 2, or 3d) This type parameter is needed since there are functions in the Polytope interface that return containers with Point{D} objects

source

InplaceArrays.ReferenceFEs.Polytope — Method.

source

InplaceArrays.ReferenceFEs.Polytope — Method.

source

InplaceArrays.ReferenceFEs.dof_cache — Method.

source

InplaceArrays.ReferenceFEs.dof_return_type — Method.

source

InplaceArrays.ReferenceFEs.edge_tangents — Method.

source

InplaceArrays.ReferenceFEs.evaluate_dof! — Method.

source

InplaceArrays.ReferenceFEs.evaluate_dof — Method.

source

InplaceArrays.ReferenceFEs.evaluate_dof_array — Method.

source

InplaceArrays.ReferenceFEs.facet_normals — Method.

source

InplaceArrays.ReferenceFEs.facet_orientations — Method.

source

InplaceArrays.ReferenceFEs.num_dims — Method.

source

InplaceArrays.ReferenceFEs.num_edges — Method.

source

InplaceArrays.ReferenceFEs.num_faces — Method.

source

InplaceArrays.ReferenceFEs.num_faces — Method.

source

InplaceArrays.ReferenceFEs.num_facets — Method.

source

InplaceArrays.ReferenceFEs.polytope_dimrange — Method.

source

InplaceArrays.ReferenceFEs.polytope_facedims — Method.

source

InplaceArrays.ReferenceFEs.polytope_faces — Method.

source

InplaceArrays.ReferenceFEs.polytope_faces — Method.

source

InplaceArrays.ReferenceFEs.polytope_offset — Method.

source

InplaceArrays.ReferenceFEs.polytope_offsets — Method.

source

InplaceArrays.ReferenceFEs.polytope_vtkid — Method.

source

InplaceArrays.ReferenceFEs.polytope_vtknodes — Method.

source

InplaceArrays.ReferenceFEs.test_dof — Function.

source

InplaceArrays.ReferenceFEs.test_polytope — Method.

It does not test polytope_vtkid nor polytope_vtknodes

source

InplaceArrays.ReferenceFEs.vertex_coordinates — Method.

source

InplaceArrays.ReferenceFEs.vertex_permutations — Method.

source

Base.:== — Method.

source

InplaceArrays.Fields.evaluate — Method.

source

InplaceArrays.Fields.evaluate — Method.

source

InplaceArrays.ReferenceFEs.num_vertices — Method.

source